Experimental features

These are experimental features. Forward/backward compatibility does not follow Julia's semantic versioning rules. Instead, compatibility is only guaranteed across changes in patch version, but not across changes of minor (or major) version changes.

BAT.bat_compare — Function

bat_compare(
    samples_1::DensitySampleVector,
    samples_2::DensitySampleVector;
    nsamples::Symbol=:effective
)

Compares two DensitySampleVectors given by samples_1 and samples_2 applying the Kolmogorov-Smirnov test for all marginals.

nsamples specifies how to define a number of samples in the Kolmogorov-Smirnov distribution. The default value is nsamples=:effective, which uses the effective number of samples estimated by bat_eff_sample_size. The optimal keywords:

:length — length of the DensitySamplesVector is used

:weights — the sum of the weights is used

Returns a NamedTuple of the shape

(result = X::TypedTables.Table, ...)

BAT.ExternalDensity — Type

BAT.ExternalDensity <: AbstractDensity

Experimental feature, not part of stable public API.

Uses an external program to calculate log-density values, the program must support the BAT binary communication protocol.

Constructor:

ExternalDensity(cmd::Cmd, density_id::Integer = 0)

BAT.FunnelDistribution — Type

BAT.FunnelDistribution([a=1, b=0.5, n=3])

Funnel distribution (Caldwell et al.)[https://arxiv.org/abs/1808.08051].

Arguments

a::Real: Variance of the dominant normal distribution.

b::Real: Variance of the supporting normal distributions.

n::Int: Number of dimensions.

Constructors:

BAT.FunnelDistribution(a::Real, b::Real, n::Int)

BAT.GaussianShell — Type

BAT.GaussianShell([r=5, w=2, n=2])

Gaussian Shell (Caldwell et al.)[https://arxiv.org/abs/1808.08051].

Arguments

r::Real: The radius of the Gaussian shell distribution.

w::Real: Variance of the Gaussian shell distribution.

n::Int: Number of dimensions.

Constructors:

BAT.GaussianShell(r::Real, w::Real, c::Vector, n::Int)

BAT.MultimodalCauchy — Type

BAT.MultimodalCauchy([μ=1, σ=0.5, n=3])

The Multimodal Cauchy Distribution (Caldwell et al.)[https://arxiv.org/abs/1808.08051].

Assumes two bimodal peaks, each in its own dimension.

Arguments

μ::Real: The location parameter used for the two bimodal peaks.

σ::Float64: The scale parameter shared among all components.

n::Int: The number of dimensions.

Constructors:

BAT.MultimodalCauchy(μ::Real, σ::Float64, n::Int)

BAT.CuhreIntegration — Type

CuhreIntegration

CuhreIntegration integration algorithm.

Only supports densities with finite rectangular bounds.

Note

This functionality is only available then the Cuba package is loaded (e.g. via import CUBA).

BAT.DivonneIntegration — Type

DivonneIntegration

DivonneIntegration integration algorithm.

Only supports densities with finite rectangular bounds.

Note

This functionality is only available then the Cuba package is loaded (e.g. via import CUBA).

BAT.GridSampler — Type

GridSampler

Constructors:

GridSampler(ppa::Int = 100)

Sample from equidistantly distributed points in each dimension.

BAT.HierarchicalDensity — Type

HierarchicalDensity <: DistLikeDensity

A hierarchical density, useful for hierarchical models/priors.

Constructors:

HierarchicalDensity(f::Function, parent_density::DistLikeDensity)

with a functon f that returns a DistLikeDensity for any variate v drawn from parent_dist.

Distributions are automatically converted to DistLikeDensitys. varshape(parent_density) and varshape(f(v)) must be a NamedTupleShape.

Example:

hd = HierarchicalDensity(
    v -> NamedTupleDist(
        baz = fill(Normal(v.bar, v.foo), 3)
    ),
    NamedTupleDist(
        foo = Exponential(3.5),
        bar = Normal(2.0, 1.0)
    )
)

varshape(hd) == NamedTupleShape(
    foo = ScalarShape{Real}(),
    bar = ScalarShape{Real}(),
    baz = ArrayShape{Real}(3)
)

v = rand(sampler(hd))
)

BAT.PartitionedSampling — Type

PartitionedSampling

BAT-internal, not part of stable public API.

The algorithm that partitions parameter space by multiple subspaces and samples/integrates them independently (See arXiv reference [not available yet]).

Constructor:

PartitionedSampling(;kwargs...)

Optional Parameters/settings (kwargs):

npartitions::Integer = 10 number of space partitions.

sampler::S = MCMCSampling() algorithm to generate samples.

exploration_sampler::S = MCMCSampling(nchains=30) algorithm to generate exploration samples.

partitioner::P = KDTreePartitioning() algorithm to partition parameter space.

integrator::I = AHMIntegration() algorithm to integrate posterior.

exploration_kwargs::NamedTuple = NamedTuple() kwargs to be used in the function bat_sample() to generate exploration samples.

sampling_kwargs::NamedTuple = NamedTuple() kwargs to be used in the function bat_sample() to generate final samples.

n_exp_samples::Integer = 10^3 number of exploration samples.

BAT.PriorImportanceSampler — Type

PriorImportanceSampler

Constructors:

PriorImportanceSampler(nsamples::Int = 10^5)

Sample randomly from prior distribution.

BAT.SobolSampler — Type

SobolSampler

Constructors:

SobolSampler(nsamples::Int = 10^5)

Sample from Sobol sequence. Also see Sobol.jl.

BAT.SuaveIntegration — Type

SuaveIntegration

SuaveIntegration integration algorithm.

Only supports densities with finite rectangular bounds.

Note

This functionality is only available then the Cuba package is loaded (e.g. via import CUBA).

BAT.VEGASIntegration — Type

VEGASIntegration

VEGASIntegration integration algorithm.

Only supports densities with finite rectangular bounds.

Note

This functionality is only available then the Cuba package is loaded (e.g. via import CUBA).