Experimental features

These are experimental features. Forward/backward compatibility does not follow Julia's semantic versioning rules. Instead, compatibility is only guaranteed across changes in patch version, but not across changes of minor (or major) version changes.

BAT.ExternalDensity — Type

BAT.ExternalDensity <: AbstractDensity

Experimental feature, not part of stable public API.

Uses an external program to calculate log-density values, the program must support the BAT binary communication protocol.

Constructor:

ExternalDensity(cmd::Cmd, density_id::Integer = 0)

BAT.FunnelDistribution — Type

BAT.FunnelDistribution([a=1, b=0.5, n=3])

Funnel distribution (Caldwell et al.)[https://arxiv.org/abs/1808.08051].

Arguments

a::Real: Variance of the dominant normal distribution.

b::Real: Variance of the supporting normal distributions.

n::Int: Number of dimensions.

Constructors:

BAT.FunnelDistribution(a::Real, b::Real, n::Int)

BAT.MultimodalCauchy — Type

BAT.MultimodalCauchy([μ=1, σ=0.5, n=3])

The Multimodal Cauchy Distribution (Caldwell et al.)[https://arxiv.org/abs/1808.08051].

Assumes two bimodal peaks, each in its own dimension.

Arguments

μ::Real: The location parameter used for the two bimodal peaks.

σ::Float64: The scale parameter shared among all components.

n::Int: The number of dimensions.

Constructors:

BAT.MultimodalCauchy(μ::Real, σ::Float64, n::Int)

BAT.GridSampler — Type

GridSampler

Constructors:

GridSampler()

Sample from equidistantly distributed points in each dimension.

BAT.HierarchicalDensity — Type

HierarchicalDensity <: DistLikeDensity

A hierarchical density, useful for hierarchical models/priors.

Constructors:

HierarchicalDensity(f::Function, parent_density::DistLikeDensity)

with a functon f that returns a DistLikeDensity for any variate v drawn from parent_dist.

Distributions are automatically converted to DistLikeDensitys. varshape(parent_density) and varshape(f(v)) must be a NamedTupleShape.

Example:

hd = HierarchicalDensity(
    v -> NamedTupleDist(
        baz = fill(Normal(v.bar, v.foo), 3)
    ),
    NamedTupleDist(
        foo = Exponential(3.5),
        bar = Normal(2.0, 1.0)
    )
)

varshape(hd) == NamedTupleShape(
    foo = ScalarShape{Real}(),
    bar = ScalarShape{Real}(),
    baz = ArrayShape{Real}(3)
)

v = rand(sampler(hd))
BAT.logvalof_unchecked(hd, v)
)

BAT.PartitionedSampling — Type

PartitionedSampling

BAT-internal, not part of stable public API.

The algorithm that partitions parameter space by multiple subspaces and samples/integrates them independently (See arXiv reference [not available yet]).

Constructor:

PartitionedSampling(;kwargs...)

Optional Parameters/settings (kwargs):

sampler::S = MetropolisHastings() algorithm to generate samples.

exploration_sampler::S = sampler algorithm to generate exploration samples.

partitioner::P = KDTreePartitioning() algorithm to partition parameter space.

integrator::I = AHMIntegration() algorithm to integrate posterior.

exploration_kwargs::NamedTuple = NamedTuple() kwargs to be used in exploration sampler.

sampling_kwargs::NamedTuple = NamedTuple() kwargs to be used in subspace sampler.

n_exploration::Tuple{Integer,Integer} = (10^2, 40) number of exploration iterations.

BAT.PriorImportanceSampler — Type

PriorImportanceSampler

Constructors:

PriorImportanceSampler()

Sample randomly from prior distribution.

BAT.SobolSampler — Type

SobolSampler

Constructors:

SobolSampler()

Sample from Sobol sequence. Also see Sobol.jl.