Experimental features

These are experimental features. Forward/backward compatibility does not follow Julia's semantic versioning rules. Instead, compatibility is only guaranteed across changes in patch version, but not across changes of minor (or major) version changes.

BAT.ExternalDensity — Type

BAT.ExternalDensity <: AbstractDensity

Experimental feature, not part of stable public API.

Uses an external program to calculate log-density values, the program must support the BAT binary communication protocol.

Constructor:

ExternalDensity(cmd::Cmd, density_id::Integer = 0)