Experimental features

These are experimental features. Forward/backward compatibility does not follow Julia's semantic versioning rules. Instead, compatibility is only guaranteed across changes in patch version, but not across changes of minor (or major) version changes.

The features listed here are likely to transition to the stable API in future versions, but may still evolve in a API-breaking fashion during that process.

BAT.bat_compare — Function

bat_compare(
    samples_1::DensitySampleVector,
    samples_2::DensitySampleVector;
    nsamples::Symbol=:effective
)

Compares two DensitySampleVectors given by samples_1 and samples_2 applying the Kolmogorov-Smirnov test for all marginals.

nsamples specifies how to define a number of samples in the Kolmogorov-Smirnov distribution. The default value is nsamples=:effective, which uses the effective number of samples estimated by bat_eff_sample_size. The optimal keywords:

:length — length of the DensitySamplesVector is used

:weights — the sum of the weights is used

Returns a NamedTuple of the shape

(result = X::TypedTables.Table, ...)

BAT.bat_rng — Function

bat_rng()

Experimental feature, not part of stable public API.

Return a new BAT-compatible random number generator, with a random seed drawn from the system entropy pool.

BAT.bat_marginalmode — Function

bat_marginalmode(
    samples::DensitySampleVector;
    nbins::Union{Integer, Symbol} = 200
)::DensitySampleVector

Experimental feature, not part of stable public API.

Estimates a local mode of samples by finding the maximum of marginalized posterior for each dimension.

Returns a NamedTuple of the shape

(result = X::DensitySampleVector, ...)

nbins specifies the number of bins that are used for marginalization. The default value is nbins=200. The optimal number of bins can be estimated using the following keywords:

:sqrt — Square-root choice

:sturges — Sturges' formula

:rice — Rice Rule

:scott — Scott's normal reference rule

:fd — Freedman–Diaconis rule

Returns a NamedTuple of the shape

(result = X::DensitySampleVector, ...)

Result properties not listed here are algorithm-specific and are not part of the stable public API.

Note

Do not add add methods to bat_marginalmode, add methods to bat_marginalmode_impl instead.

BAT.ExternalDensity — Type

struct BAT.ExternalDensity <: AbstractDensity

Experimental feature, not part of stable public API.

Uses an external program to calculate log-density values, the program must support the BAT binary communication protocol.

Constructor:

ExternalDensity(cmd::Cmd, density_id::Integer = 0)

All fields of this type are considered internal and subject to change without deprecation.

BAT.FunnelDistribution — Type

struct BAT.FunnelDistribution <: Distribution{Multivariate,Continuous}

Experimental feature, not part of stable public API.

A funnel distribution (see Caldwell et al. for definition).

Constructors:

FunnelDistribution(; a::Real = 1.0, b::Real = 0.5, n::Integer = 3)

Fields:

a::Real

Variance of the dominant normal distribution.

b::Real

Variance of the supporting normal distributions.

n::Integer

Number of dimensions.

BAT.GaussianShell — Type

struct BAT.GaussianShell <: Distribution{Multivariate,Continuous}

Experimental feature, not part of stable public API.

Gaussian Shell (see Caldwell et al. for definition).

Constructors:

BAT.GaussianShell(; r::Real=5, w::Real=2, n::Integer=2)

Fields:

r::Real

The radius of the Gaussian shell distribution.

w::Real

Variance of the Gaussian shell distribution

n::Int64

Number of dimensions

c::AbstractArray{var"#s48",1} where var"#s48"<:Real

lognorm::AbstractFloat

Note

Fields c and lognorm are considered internal and subject to change without deprecation.

BAT.MultimodalCauchy — Type

BAT.MultimodalCauchy <: Distribution{Multivariate,Continuous}

Experimental feature, not part of stable public API.

The Multimodal Cauchy Distribution (see Caldwell et al. for definition).

Assumes two bimodal peaks, each in its own dimension.

Constructors:

BAT.MultimodalCauchy(; μ::Real=1, σ::Float64=0.2, n::Integer=4)

Arguments:

μ::Real: The location parameter used for the two bimodal peaks.

σ::Float64: The scale parameter shared among all components.

n::Int: The number of dimensions.

Fields:

bimodals::Distributions.MixtureModel

σ::Float64

n::Int64

dist::Distributions.Product

Note

All fields of MultimodalCauchy are considered internal and subject to change without deprecation.

BAT.LogUniform — Type

struct BAT.LogUniform{T<:Real} <: Distributions.Distribution{Univariate,Continuous}

Experimental feature, not part of stable public API.

The log-uniform distribution (reciprocal distribution) over an interval $[a, b]$.

Constructors:

LogUniform(a::Real, b::Real)

See also Reciprocal distribution on Wikipedia.

BAT.CuhreIntegration — Type

struct CuhreIntegration <: IntegrationAlgorithm

Experimental feature, not part of stable public API.

CuhreIntegration integration algorithm.

Constructors:

CuhreIntegration(; fields...)

Fields:

trafo::AbstractDensityTransformTarget

Default: PriorToUniform()

log_density_shift::Float64

Default: 0.0

rtol::Float64

Default: Cuba.RTOL

atol::Float64

Default: Cuba.ATOL

minevals::Int64

Default: Cuba.MINEVALS

maxevals::Int64

Default: Cuba.MAXEVALS

key::Int64

Default: Cuba.KEY

nthreads::Int64

Default: Base.Threads.nthreads()

strict::Bool

Default: true

Note

This functionality is only available when the Cuba package is loaded (e.g. via import CUBA).

BAT.DivonneIntegration — Type

struct DivonneIntegration <: IntegrationAlgorithm

Experimental feature, not part of stable public API.

DivonneIntegration integration algorithm.

Constructors:

DivonneIntegration(; fields...)

Fields:

trafo::AbstractDensityTransformTarget

Default: PriorToUniform()

log_density_shift::Float64

Default: 0.0

rtol::Float64

Default: Cuba.RTOL

atol::Float64

Default: Cuba.ATOL

minevals::Int64

Default: Cuba.MINEVALS

maxevals::Int64

Default: Cuba.MAXEVALS

key1::Int64

Default: Cuba.KEY1

key2::Int64

Default: Cuba.KEY2

key3::Int64

Default: Cuba.KEY3

maxpass::Int64

Default: Cuba.MAXPASS

border::Float64

Default: Cuba.BORDER

maxchisq::Float64

Default: Cuba.MAXCHISQ

mindeviation::Float64

Default: Cuba.MINDEVIATION

ngiven::Int64

Default: Cuba.NGIVEN

ldxgiven::Int64

Default: Cuba.LDXGIVEN

nextra::Int64

Default: Cuba.NEXTRA

nthreads::Int64

Default: Base.Threads.nthreads()

strict::Bool

Default: true

Note

This functionality is only available when the Cuba package is loaded (e.g. via import CUBA).

BAT.GridSampler — Type

struct GridSampler <: AbstractSamplingAlgorithm

Experimental feature, not part of stable public API.

Sample from equidistantly distributed points in each dimension.

Constructors:

GridSampler(; fields...)

Fields:

trafo::AbstractDensityTransformTarget

Default: PriorToUniform()

ppa::Int64

Default: 100

BAT.HierarchicalDistribution — Type

struct HierarchicalDistribution <: ContinuousDistribution

Experimental feature, not part of stable public API.

A hierarchical distribution, useful for hierarchical models/priors.

Constructors:

HierarchicalDistribution(f::Function, primary_dist::NamedTupleDist)

with a functon f that returns a ContinuousDistribution for any variate v drawn from primary_dist.

Example:

hd = HierarchicalDistribution(
    v -> NamedTupleDist(
        baz = fill(Normal(v.bar, v.foo), 3)
    ),
    NamedTupleDist(
        foo = Exponential(3.5),
        bar = Normal(2.0, 1.0)
    )
)

varshape(hd) == NamedTupleShape(
    foo = ScalarShape{Real}(),
    bar = ScalarShape{Real}(),
    baz = ArrayShape{Real}(3)
)

v = rand(hd)

Note

All fields of HierarchicalDistribution are considered internal and subject to change without deprecation.

BAT.PartitionedSampling — Type

struct PartitionedSampling <: AbstractSamplingAlgorithm

Experimental feature, not part of stable public API.

A sampling algorithm that partitions parameter space into multiple subspaces and samples/integrates them independently (Caldwell et al.).

Constructors:

PartitionedSampling(; fields...)

Fields:

trafo::AbstractDensityTransformTarget

Default: PriorToUniform()

npartitions::Integer

Default: 10

sampler::BAT.AbstractSamplingAlgorithm

Default: MCMCSampling()

exploration_sampler::BAT.AbstractSamplingAlgorithm

Default: MCMCSampling(nchains = 30, nsteps = 1000, trafo = NoDensityTransform())

partitioner::SpacePartitioningAlgorithm

Default: KDTreePartitioning()

integrator::IntegrationAlgorithm

Default: AHMIntegration()

BAT.PriorImportanceSampler — Type

struct PriorImportanceSampler <: AbstractSamplingAlgorithm

Experimental feature, not part of stable public API.

Importance sampler using IID samples from the prior.

Constructors:

PriorImportanceSampler(; fields...)

Fields:

nsamples::Int64

Default: 10 ^ 5

BAT.ReactiveNestedSampling — Type

struct ReactiveNestedSampling <: AbstractUltraNestAlgorithm

Experimental feature, not part of stable public API.

UltraNest reactive nested sampling algorithm with.

Uses the UltraNest Python package, via UltraNest.jl (and PyCall).

Constructors:

ReactiveNestedSampling(; fields...)

Fields:

trafo::AbstractDensityTransformTarget

Default: PriorToUniform()

num_test_samples::Int64

Test transform and likelihood with this number of random points for errors first. Useful to catch bugs. Default: 2

draw_multiple::Bool

If efficiency goes down, dynamically draw more points from the region between ndrawmin and ndrawmax. If set to False, few points are sampled at once. Default: true

num_bootstraps::Int64

Number of logZ estimators and MLFriends region bootstrap rounds. Default: 30

ndraw_min::Int64

Minimum number of points to simultaneously propose. Increase this if your likelihood makes vectorization very cheap. Default: 128

ndraw_max::Int64

Maximum number of points to simultaneously propose. Increase this if your likelihood makes vectorization very cheap. Memory allocation may be slow for extremely high values. Default: 65536

update_interval_volume_fraction::Float64

Update region when the volume shrunk by this amount. Default: 0.8

log_interval::Int64

Update stdout status line every log_interval iterations. Default: -1

show_status::Bool

Show integration progress as a status line. Default: true

dlogz::Float64

Target evidence uncertainty. This is the std between bootstrapped logz integrators. Default: 0.5

dKL::Float64

Target posterior uncertainty. This is the Kullback-Leibler divergence in nat between bootstrapped integrators. Default: 0.5

frac_remain::Float64

Integrate until this fraction of the integral is left in the remainder. Set to a low number (1e-2 … 1e-5) to make sure peaks are discovered. Set to a higher number (0.5) if you know the posterior is simple. Default: 0.01

Lepsilon::Float64

Terminate when live point likelihoods are all the same, within Lepsilon tolerance. Increase this when your likelihood function is inaccurate, to avoid unnecessary search. Default: 0.001

min_ess::Int64

Target number of effective posterior samples. Default: 400

max_iters::Int64

maximum number of integration iterations. Default: -1

max_ncalls::Int64

Stop after this many likelihood evaluations. Default: -1

max_num_improvement_loops::Int64

The algorithm tries to assess iteratively where more samples are needed. This number limits the number of improvement loops. Default: -1

min_num_live_points::Int64

Minimum number of live points throughout the run. Default: 400

cluster_num_live_points::Int64

Require at least this many live points per detected cluster. Default: 40

insertion_test_window::Float64

z-score used as a threshold for the insertion order test. Set to infinity to disable. Default: 10.0

insertion_test_zscore_threshold::Float64

Number of iterations after which the insertion order test is reset. Default: 2.0

Note

This functionality is only available when the UltraNest package is loaded (e.g. via import UltraNest).

BAT.SobolSampler — Type

struct SobolSampler <: AbstractSamplingAlgorithm

Experimental feature, not part of stable public API.

Sample from Sobol sequence. Also see Sobol.jl.

Constructors:

SobolSampler(; fields...)

Fields:

trafo::AbstractDensityTransformTarget

Default: PriorToUniform()

nsamples::Int64

Default: 10 ^ 5

BAT.SuaveIntegration — Type

struct SuaveIntegration <: IntegrationAlgorithm

Experimental feature, not part of stable public API.

SuaveIntegration integration algorithm.

Constructors:

SuaveIntegration(; fields...)

Fields:

trafo::AbstractDensityTransformTarget

Default: PriorToUniform()

log_density_shift::Float64

Default: 0.0

rtol::Float64

Default: Cuba.RTOL

atol::Float64

Default: Cuba.ATOL

minevals::Int64

Default: Cuba.MINEVALS

maxevals::Int64

Default: Cuba.MAXEVALS

nnew::Int64

Default: Cuba.NNEW

nmin::Int64

Default: Cuba.NMIN

flatness::Float64

Default: Cuba.FLATNESS

nthreads::Int64

Default: Base.Threads.nthreads()

strict::Bool

Default: true

Note

This functionality is only available when the Cuba package is loaded (e.g. via import CUBA).

BAT.VEGASIntegration — Type

struct VEGASIntegration <: IntegrationAlgorithm

Experimental feature, not part of stable public API.

VEGASIntegration integration algorithm.

Constructors:

VEGASIntegration(; fields...)

Fields:

trafo::AbstractDensityTransformTarget

Default: PriorToUniform()

log_density_shift::Float64

Default: 0.0

rtol::Float64

Default: Cuba.RTOL

atol::Float64

Default: Cuba.ATOL

minevals::Int64

Default: Cuba.MINEVALS

maxevals::Int64

Default: Cuba.MAXEVALS

nstart::Int64

Default: Cuba.NSTART

nincrease::Int64

Default: Cuba.NINCREASE

nbatch::Int64

Default: Cuba.NBATCH

nthreads::Int64

Default: Base.Threads.nthreads()

strict::Bool

Default: true

Note

This functionality is only available when the Cuba package is loaded (e.g. via import CUBA).