Internal API

Note

This is the documentation of BAT's internal API. The internal API is fully accessible to users, but all aspects of it are subject to change without deprecation. Functionalities of the internal API that, over time, turn out to require user access (e.g. to support advanced use cases) will be evolved to gain a stable interface and then promoted to the public API.

Types

BAT.AbstractProposalDist

BAT.AbstractSampleGenerator

BAT.BasicMvStatistics

BAT.ENSAutoProposal

BAT.ENSBound

BAT.ENSEllipsoidBound

BAT.ENSMultiEllipsoidBound

BAT.ENSNoBounds

BAT.ENSProposal

BAT.ENSRandomWalk

BAT.ENSSlice

BAT.ENSUniformly

BAT.LFDensity

BAT.LFDensityWithGrad

BAT.MCMCSampleGenerator

BAT.OnlineMvCov

BAT.OnlineMvMean

BAT.OnlineUvMean

BAT.OnlineUvVar

BAT.RenormalizedDensity

BAT.StandardMvNormal

BAT.StandardMvUniform

BAT.StandardUvNormal

BAT.StandardUvUniform

BAT.TransformedDensity

BAT.WrappedNonBATDensity

Functions and macros

BAT.argchoice_msg

BAT.bat_sampler

BAT.bg_R_2sqr

BAT.checked_logdensityof

BAT.default_val_numtype

BAT.default_var_numtype

BAT.density_valtype

BAT.drop_low_weight_samples

BAT.find_marginalmodes

BAT.fromuhc

BAT.fromuhc!

BAT.fromui

BAT.get_bin_centers

BAT.getlikelihood

BAT.getprior

BAT.gr_Rsqr

BAT.is_log_zero

BAT.issymmetric_around_origin

BAT.log_volume

BAT.log_zero_density

BAT.proposal_rand!

BAT.proposaldist_logpdf

BAT.repetition_to_weights

BAT.spatialvolume

BAT.sum_first_dim

BAT.trunc_logpdf_ratio

BAT.truncate_dist_hard

BAT.var_bounds

Documentation

BAT.AbstractProposalDist — Type

abstract type AbstractProposalDist

BAT-internal, not part of stable public API.

The following functions must be implemented for subtypes:

BAT.proposaldist_logpdf

BAT.proposal_rand!

ValueShapes.totalndof, returning the number of DOF (i.e. dimensionality).

LinearAlgebra.issymmetric, indicating whether p(a -> b) == p(b -> a) holds true.

BAT.AbstractSampleGenerator — Type

abstract type AbstractSampleGenerator

BAT-internal, not part of stable public API.

Abstract super type for sample generators.

BAT.BasicMvStatistics — Type

BasicMvStatistics{T<:Real,W}

BAT-internal, not part of stable public API.

W must either be Weights (no bias correction) or one of AnalyticWeights, FrequencyWeights or ProbabilityWeights to specify the desired bias correction method.

BAT.ENSAutoProposal — Type

struct BAT.ENSAutoProposal <: ENSProposal

Experimental feature, not part of stable public API.

Choose the proposal depending from the number of dimensions: ndims < 10: Proposals.Uniform, 10 ≤ ndims ≤ 20: Proposals.RWalk, ndims > 20: Proposals.Slice

BAT.ENSBound — Type

abstract type ENSBound

Experimental feature, not part of stable public API.

Abstract type for the bounds of the sampling region used by EllipsoidalNestedSampling.

BAT.ENSEllipsoidBound — Type

struct BAT.ENSEllipsoidBound <: ENSBound

Experimental feature, not part of stable public API.

Ellipsoid bound means that a n-dimensional ellipsoid limits the sampling volume.

BAT.ENSMultiEllipsoidBound — Type

struct BAT.ENSMultiEllipsoidBound <: ENSBound

Experimental feature, not part of stable public API.

Multi ellipsoid bound means that there are multiple elliposid in an optimal clustering are used to limit the sampling volume.

BAT.ENSNoBounds — Type

struct BAT.ENSNoBounds <: ENSBound

Experimental feature, not part of stable public API.

No bounds means that the whole volume from the unit Cube is used to find new points.

BAT.ENSProposal — Type

abstract type ENSProposal

Experimental feature, not part of stable public API.

Abstract type for the algorithms to propose new live points used used by EllipsoidalNestedSampling.

BAT.ENSRandomWalk — Type

struct BAT.ENSRandomWalk <: ENSProposal

Experimental feature, not part of stable public API.

New live point is proposed by using a random walk away from an existing live point.

Constructors:

ENSRandomWalk(; fields...)

Fields:

ratio::Float64

Acceptance ratio for the random walk. Default: 0.5

walks::Int64

Minimum number of random walk steps. Default: 25

scale::Float64

Scale of the proposal distribution. Default: 1.0

BAT.ENSSlice — Type

struct BAT.ENSSlice <: ENSProposal

Experimental feature, not part of stable public API.

New live point is proposed by a serie of random slices from an existing live-point.

Constructors:

ENSSlice(; fields...)

Fields:

slices::Int64

Minimum number of slices Default: 5

scale::Float64

Scale of the proposal distribution. Default: 1.0

BAT.ENSUniformly — Type

struct BAT.ENSUniformly <: ENSProposal

Experimental feature, not part of stable public API.

Each point in the bounding volume has an uniform chance to be proposed as a new live point.

BAT.LFDensity — Type

struct BAT.LFDensity{F}

BAT-internal, not part of stable public API.

Wraps a log-density function log_f.

BAT.LFDensityWithGrad — Type

BAT.LFDensityWithGrad{F<:Function} <: AbstractDensity

BAT-internal, not part of stable public API.

Constructors:

LFDensityWithGrad(logf::Function, valgradlogf::Function)

A density defined by a function that computes it's logarithmic value at given points, as well as a function that computes both the value and the gradient.

It must be safe to execute both functions in parallel on multiple threads and processes.

BAT.MCMCSampleGenerator — Type

BAT.MCMCSampleGenerator

BAT-internal, not part of stable public API.

MCMC sample generator.

Constructors:

MCMCSampleGenerator(chain::AbstractVector{<:MCMCIterator})

BAT.OnlineMvCov — Type

OnlineMvCov{T<:AbstractFloat,W} <: AbstractMatrix{T}

BAT-internal, not part of stable public API.

Implementation based on variance calculation Algorithms of Welford and West.

W must either be Weights (no bias correction) or one of AnalyticWeights, FrequencyWeights or ProbabilityWeights to specify the desired bias correction method.

BAT.OnlineMvMean — Type

OnlineMvMean{T<:AbstractFloat} <: AbstractVector{T}

BAT-internal, not part of stable public API.

Multivariate mean implemented via Kahan-Babuška-Neumaier summation.

BAT.OnlineUvMean — Type

OnlineUvMean{T<:AbstractFloat}

BAT-internal, not part of stable public API.

Univariate mean implemented via Kahan-Babuška-Neumaier summation.

BAT.OnlineUvVar — Type

OnlineUvVar{T<:AbstractFloat,W}

BAT-internal, not part of stable public API.

Implementation based on variance calculation Algorithms of Welford and West.

W must either be Weights (no bias correction) or one of AnalyticWeights, FrequencyWeights or ProbabilityWeights to specify the desired bias correction method.

BAT.RenormalizedDensity — Type

struct RenormalizedDensity <: AbstractDensity

Constructors:

RenormalizedDensity(density::AbstractDensity, logrenormf::Real)

A renormalized density derived from density, with

logdensityof(RenormalizedDensity(density, logrenormf))(v) ==
    logdensityof(density)(v) + logrenormf

BAT.StandardMvNormal — Type

StandardMvNormal{T<:Real} <: Distributions.AbstractMvNormal

A standard n-dimensional multivariate normal distribution with it's mean at the origin and an identity covariance matrix.

Constructor:

    StandardMvNormal(n::Integer)
    StandardMvNormal{T<:Real}(n::Integer)

BAT.StandardMvUniform — Type

StandardMvUniform{T<:Real} <: Distributions.Distribution{Multivariate,Continuous}

A standard n-dimensional multivariate uniform distribution, from zero to one in each dimension.

Constructor:

    StandardMvUniform(n::Integer)
    StandardMvUniform{T<:Real}(n::Integer)

BAT.StandardUvNormal — Type

StandardUvNormal{T<:Real} <: Distributions.Distribution{Univariate,Continuous}

A standard normal distribution with a mean of zero and a variance of one.

Constructor:

    StandardUvNormal()
    StandardUvNormal{T<:Real}()

BAT.StandardUvUniform — Type

StandardUvUniform{T<:Real} <: Distributions.Distribution{Univariate,Continuous}

A standard uniform distribution between zero and one.

Constructor:

    StandardUvUniform()
    StandardUvUniform{T<:Real}()

BAT.TransformedDensity — Type

TransformedDensity

BAT-internal, not part of stable public API.

BAT.WrappedNonBATDensity — Type

struct BAT.WrappedNonBATDensity{F<:Base.Callable}

BAT-internal, not part of stable public API.

Wraps a log-density function log_f.

BAT.argchoice_msg — Function

argchoice_msg(f::Base.Callable, argname::Val, x)

BAT-internal, not part of stable public API.

Generates an information message regarding the choice of value x for argument argname of function f.

The value x will often be the result of bat_default.

BAT.bat_sampler — Function

bat_sampler(d::Distribution)

BAT-internal, not part of stable public API.

Tries to return a BAT-compatible sampler for Distribution d. A sampler is BAT-compatible if it supports random number generation using an arbitrary AbstractRNG:

rand(rng::AbstractRNG, s::SamplerType)
rand!(rng::AbstractRNG, s::SamplerType, x::AbstractArray)

If no specific method of bat_sampler is defined for the type of d, it will default to sampler(d), which may or may not return a BAT-compatible sampler.

BAT.bg_R_2sqr — Function

bg_R_2sqr(stats::AbstractVector{<:MCMCBasicStats}; corrected::Bool = false)
bg_R_2sqr(samples::AbstractVector{<:DensitySampleVector}; corrected::Bool = false)

BAT-internal, not part of stable public API.

Brooks-Gelman R_2^2 for all DOF. If normality is assumed, 'corrected' should be set to true to account for the sampling variability.

BAT.checked_logdensityof — Function

checked_logdensityof(density::AbstractDensity, v::Any, T::Type{<:Real})

BAT-internal, not part of stable public API.

Evaluates density log-value via DensityInterface.logdensityof and performs additional checks.

Throws a BAT.DensityEvalException on any of these conditions:

The variate shape of density (if known) does not match the shape of v.

The return value of DensityInterface.logdensityof is NaN.

The return value of DensityInterface.logdensityof is an equivalent of positive infinity.

BAT.default_val_numtype — Function

BAT.default_val_numtype(density::AbstractDensity)

BAT-internal, not part of stable public API.

Returns the default/preferred numerical type (log-)density values of density.

BAT.default_var_numtype — Function

BAT.default_var_numtype(density::AbstractDensity)

BAT-internal, not part of stable public API.

Returns the default/preferred underlying numerical type for (elements of) variates of density.

BAT.density_valtype — Function

BAT.density_valtype(density::AbstractDensity, v::Any)

BAT-internal, not part of stable public API.

Determine a suitable return type for the (log-)density value of the given density for the given variate.

BAT.drop_low_weight_samples — Function

drop_low_weight_samples(
    samples::DensitySampleVector,
    fraction::Real = 10^-4
)

BAT-internal, not part of stable public API.

Drop fraction of the total probability mass from samples to filter out the samples with the lowest weight.

BAT.find_marginalmodes — Function

find_marginalmodes(marg::MarginalDist)

BAT-internal, not part of stable public API.

Find the modes of a MarginalDist. Returns a vector of the bin-centers of the bin(s) with the heighest weight.

BAT.fromuhc — Function

fromuhc(X::AbstractVector, vol::SpatialVolume)
fromuhc(X::VectorOfSimilarVectors, vol::SpatialVolume)

BAT-internal, not part of stable public API.

Bijective transformation from unit hypercube to vol. See fromuhc!.

Use inv(fromuhc) to get the the inverse transformation.

BAT.fromuhc! — Function

fromuhc!(Y::AbstractVector, X::AbstractVector, vol::SpatialVolume)
fromuhc!(Y::VectorOfSimilarVectors, X::VectorOfSimilarVectors, vol::SpatialVolume)

BAT-internal, not part of stable public API.

Bijective transformation of coordinates X within the unit hypercube to coordinates Y in vol. If X and Y are matrices, the transformation is applied to the column vectors. Use Y === X to transform in-place.

Use inv(fromuhc!) to get the the inverse transformation.

BAT.fromui — Function

y = fromui(x::Real, lo::Real, hi::Real)
y = fromui(x::Real, lo_hi::ClosedInterval{<:Real})

BAT-internal, not part of stable public API.

Linear bijective transformation from the unit inverval (i.e. x ∈ 0..1) to y ∈ lo..hi.

Use inv(fromui) to get the the inverse transformation.

Use @inbounds to disable range checking on the input value.

BAT.get_bin_centers — Function

get_bin_centers(marg::MarginalDist)

BAT-internal, not part of stable public API.

Returns a vector of the bin-centers.

BAT.getlikelihood — Function

getlikelihood(posterior::AbstractPosteriorDensity)::AbstractDensity

BAT-internal, not part of stable public API.

The likelihood density of posterior. The likelihood may or may not be normalized.

BAT.getprior — Function

getprior(posterior::AbstractPosteriorDensity)::AbstractDensity

BAT-internal, not part of stable public API.

The prior density of posterior. The prior may or may not be normalized.

BAT.gr_Rsqr — Function

gr_Rsqr(stats::AbstractVector{<:MCMCBasicStats})
gr_Rsqr(samples::AbstractVector{<:DensitySampleVector})

BAT-internal, not part of stable public API.

Gelman-Rubin $$R^2$$ for all DOF.

BAT.is_log_zero — Function

BAT.is_log_zero(x::Real, T::Type{<:Real} = typeof(x)}

BAT-internal, not part of stable public API.

Check if x is an equivalent of log of zero, resp. negative infinity, in respect to type T.

BAT.issymmetric_around_origin — Function

issymmetric_around_origin(d::Distribution)

BAT-internal, not part of stable public API.

Returns true (resp. false) if the Distribution is symmetric (resp. non-symmetric) around the origin.

BAT.log_volume — Function

log_volume(vol::SpatialVolume)

BAT-internal, not part of stable public API.

Get the logarithm of the volume of the space in vol.

BAT.log_zero_density — Function

BAT.log_zero_density(T::Type{<:Real})

log-density value to assume for regions of implicit zero density, e.g. outside of variate/parameter bounds/support.

Returns an equivalent of negative infinity.

BAT.proposal_rand! — Function

function proposal_rand!(
    rng::AbstractRNG,
    pdist::GenericProposalDist,
    v_proposed::Union{AbstractVector,VectorOfSimilarVectors},
    v_current::Union{AbstractVector,VectorOfSimilarVectors}
)

BAT-internal, not part of stable public API.

Generate one or multiple proposed variate/parameter vectors, based on one or multiple previous vectors.

Input:

rng: Random number generator to use

pdist: Proposal distribution to use

v_current: Old values (vector or column vectors, if a matrix)

Output is stored in

v_proposed: New values (vector or column vectors, if a matrix)

The caller must guarantee:

size(v_current, 1) == size(v_proposed, 1)

size(v_current, 2) == size(v_proposed, 2) or size(v_current, 2) == 1

v_proposed !== v_current (no aliasing)

Implementations of proposal_rand! must be thread-safe.

BAT.proposaldist_logpdf — Function

proposaldist_logpdf(
    p::AbstractArray,
    pdist::AbstractProposalDist,
    v_proposed::AbstractVector,
    v_current:::AbstractVector
)

BAT-internal, not part of stable public API.

Returns log(PDF) value of pdist for transitioning from current to proposed variate/parameters.

BAT.repetition_to_weights — Function

repetition_to_weights(v::AbstractVector)

BAT-internal, not part of stable public API.

Drop (subsequently) repeated samples by adding weights.

BAT.spatialvolume — Function

spatialvolume(b::VarVolumeBounds)::SpatialVolume

BAT-internal, not part of stable public API.

Returns the spatial volume that defines the variate/parameter bounds.

BAT.sum_first_dim — Function

@propagate_inbounds sum_first_dim(A::AbstractArray, j::Integer, ks::Integer...)

BAT-internal, not part of stable public API.

Calculate the equivalent of sum(A[:, j, ks...]).

@propagate_inbounds sum_first_dim(A::AbstractArray)

BAT-internal, not part of stable public API.

If A is a vector, return sum(A), else sum(A, 1)[:].

BAT.trunc_logpdf_ratio — Function

BAT.trunc_logpdf_ratio(orig_dist::Distribution{TP}, trunc_dist::Distribution{TP})::AbstractFloat

BAT-internal, not part of stable public API.

Computes the log-ratio between the amplitude of the PDF of a truncated distribution and the original (untruncted) distribution, within the support of the truncated one.

The PDF of both distributions must have the same shape within the support of trunc_dist and may only differ in amplitude.

Mainly used to implement BAT.truncate_density, in conjunction with BAT.truncate_dist_hard. The result contributes to the logrenormf factor of a [RenormalizedDensity] that uses truncated distributions internally, to ensure the density does not get renormalized.

BAT.truncate_dist_hard — Function

BAT.truncate_dist_hard(dist::Distribution{Univariate}, bounds::Interval)::Distribution{Univariate}
BAT.truncate_dist_hard(dist::Distribution{Multivariate}, bounds::AbstractArray{<:Interval})::Distribution{Multivariate}

BAT-internal, not part of stable public API.

Generalized variant of Distributions.truncated - also handles multivariate distributions and operates on a best-effort basis: If distributions cannot be truncated, may return the original distribution.

Returns a NamedTuple

    (dist = trunc_dist, logrenormf = logrenormf)

with the truncated distribution and the log-PDF amplitude difference to the original (see BAT.trunc_logpdf_ratio).

Mainly used to implement BAT.truncate_density.

BAT.var_bounds — Function

var_bounds(density::DistLikeDensity)::AbstractVarBounds

BAT-internal, not part of stable public API.

Get the parameter bounds of density. Must not be missing.