Internal API

Note

This is the documentation of BAT's internal API. The internal API is fully accessible to users, but all aspects of it are subject to change without notice. Functionalities of the internal API that, over time, turn out to require user access (e.g. to support advanced use cases) will be evolved to gain a stable interface and then promoted to the public API.

Types

BAT.AbstractMCMCCallback

BAT.AbstractProposalDist

BAT.BasicMvStatistics

BAT.DataSet

BAT.HMIData

BAT.HMISettings

BAT.IntegrationVolume

BAT.MCMCCallbackWrapper

BAT.MCMCSpec

BAT.OnlineMvCov

BAT.OnlineMvMean

BAT.OnlineUvMean

BAT.OnlineUvVar

BAT.PointCloud

BAT.SearchResult

BAT.WhiteningResult

Functions and macros

BAT.apply_bounds

BAT.apply_bounds!

BAT.apply_bounds_and_eval_posterior_logval!

BAT.apply_bounds_and_eval_posterior_logval_strict!

BAT.autocrl

BAT.bat_sampler

BAT.bg_R_2sqr

BAT.calculate_localmode

BAT.create_hypercube

BAT.create_hyperrectangle

BAT.distribution_logpdf

BAT.distribution_logpdf!

BAT.drop_low_weight_samples

BAT.effective_sample_size

BAT.eval_density_logval

BAT.find_hypercube_centers

BAT.fromuhc

BAT.fromuhc!

BAT.fromui

BAT.getlikelihood

BAT.getprior

BAT.gr_Rsqr

BAT.hm_init

BAT.hm_integrate!

BAT.hm_whiteningtransformation!

BAT.hyperrectangle_creationproccess!

BAT.integrate_hyperrectangle_cov

BAT.issymmetric_around_origin

BAT.log_volume

BAT.mcmc_startval!

BAT.modify_hypercube!

BAT.modify_integrationvolume!

BAT.proposal_rand!

BAT.reduced_volume_hm

BAT.spatialvolume

BAT.sum_first_dim

BAT.var_bounds

BAT.wgt_effective_sample_size

Documentation

BAT.AbstractMCMCCallback — Type

AbstractMCMCCallback <: Function

BAT-internal, not part of stable public API.

Subtypes (e.g. X) must support

(::X)(level::Integer, chain::MCMCIterator) => nothing
(::X)(level::Integer, tuner::AbstractMCMCTuner) => nothing

to be compabtible with mcmc_iterate!, mcmc_tune_burnin!, etc.

BAT.AbstractProposalDist — Type

AbstractProposalDist

BAT-internal, not part of stable public API.

The following functions must be implemented for subtypes:

BAT.distribution_logpdf

BAT.proposal_rand!

ValueShapes.totalndof, returning the number of DOF (i.e. dimensionality).

LinearAlgebra.issymmetric, indicating whether p(a -> b) == p(b -> a) holds true.

In some cases, it may be desirable to override the default implementation of BAT.distribution_logpdf!.

BAT.BasicMvStatistics — Type

BasicMvStatistics{T<:Real,W}

BAT-internal, not part of stable public API.

W must either be Weights (no bias correction) or one of AnalyticWeights, FrequencyWeights or ProbabilityWeights to specify the desired bias correction method.

BAT.DataSet — Type

DataSet{T<:AbstractFloat, I<:Integer}

AHMI-internal, not part of stable public API.

Holds the MCMC output. For construction use constructor: function DataSet{T<:Real}(data::Matrix{T}, logprob::Vector{T}, weights::Vector{T})

Variables

'data' : An P x N array with N data points with P parameters.

'logprob' : The logarithmic probability for each samples stored in an array

'weights' : How often each sample occurred. Set to an array of ones if working directly on MCMC output

'ids' : Array which is used to assign each sample to a batch, required for the cov. weighed uncertainty estimation

.sortids : an array of indices which stores the original ordering of the samples (the space partitioning tree reorders the samples), required to calculate an effective sample size.

'N' : number of samples

'P' : number of parameters

'nsubsets' : the number of batches

'iswhitened' : a boolean value which indicates whether the data set is iswhitened

'isnew' : a boolean value which indicates whether the data set was swapped out with a different one (it is possible to redo the integration with a different sample set using previously generated hyper-rectangles)

'partitioningtree' : The space partitioning tree, used to efficiently identify samples in a point cloud

'startingIDs' : The Hyper-Rectangle Seed Samples are stored in this array

'tolerance' : A threshold required for the hyper-rectangle creation process.

BAT.HMIData — Type

HMIData{T<:AbstractFloat, I<:Integer}

AHMI-internal, not part of stable public API.

Includes all the informations of the integration process, including a list of hyper-rectangles, the results of the whitening transformation, the starting ids, and the average number of points and volume of the created hyper-rectangles.

Variables

'dataset1' : Data Set 1

'dataset2' : Data Set 2

'whiteningresult' : contains the whitening matrix and its determinant, required to scale the final integral estimate

'volumelist1' : An array of integration volumes created using dataset1, but filled with samples from dataset2

'volumelist2' : An array of integration volumes created using dataset2, but filled with samples from dataset1

'cubelist1' : An array of small hyper-cubes created around seeding samples of dataset 1

'cubelist2' : An array of small hyper-cubes created around seeding samples of dataset 2

'iterations1' : The number of volume adapting iterations for the creating volumelist1

'iterations2' : The number of volume adapting iterations for the creating volumelist2

'rejectedrects1' : An array of ids, indicating which hyper-rectangles of volumelist1 were rejected due to trimming

'rejectedrects2' : An array of ids, indicating which hyper-rectangles of volumelist2 were rejected due to trimming

'integralestimates' : A dictionary containing the final integral estimates with uncertainty estimation using different uncertainty estimators. Also includes all intermediate results required for the integral estimate combination

BAT.HMISettings — Type

HMISettings

AHMI-internal, not part of stable public API.

holds the settings for the hm_integrate function. There are several default constructors available: HMIFastSettings() HMIStandardSettings() HMIPrecisionSettings()

#Variables

'whitening_method::Symbol' : which whitening method to use

'max_startingIDs::Integer' : influences how many starting ids are allowed to be generated

'maxstartingIDsfraction::AbstractFloat' : how many points are considered as possible starting points as a fraction of total points available

'rect_increase::AbstractFloat' : describes the procentual rectangle volume increase/decrease during hyperrectangle creation. Low values can increase the precision if enough points are available but can cause systematically wrong results if not enough points are available.

'useallrects::Bool' : All rectangles are used for the integration process no matter how big their overlap is. If enabled the rectangles are weighted by their overlap.

'useMultiThreading' : activate multithreading support.

'warning_minstartingids' : the required minimum amount of starting samples

'dotrimming' : determines whether the integral estimates are trimmed (1σ trim) before combining them into a final result (more robust)

'uncertaintyestimators' : A dictionary of different uncertainty estimator functions. Currently three functions are available: hmcombineresultslegacy! (outdated, overestimates uncertainty significantly in higher dimensions), hmcombineresultscovweighted! (very fast) and hmcombineresults_analyticestimation! (recommended)

end

BAT.IntegrationVolume — Type

IntegrationVolume{T<:AbstractFloat, I<:Integer}

AHMI-internal, not part of stable public API.

Variables

'pointcloud' : holds the point cloud of the integration volume

'spatialvolume' : the boundaries of the integration volume

'volume' : the volume

Hold the point cloud and the spatial volume for integration.

BAT.MCMCCallbackWrapper — Type

MCMCCallbackWrapper{F} <: AbstractMCMCCallback

BAT-internal, not part of stable public API.

Wraps a callable object to turn it into an AbstractMCMCCallback.

Constructor:

MCMCCallbackWrapper(f::Any)

f needs to support the call syntax of an AbstractMCMCCallback.

BAT.MCMCSpec — Type

MCMCSpec{
    A<:MCMCAlgorithm,
    M<:AbstractPosteriorDensity
}

BAT-internal, not part of stable public API.

Specifies a Bayesian MCMC chain.

Constructor:

MCMCSpec(
    algorithm::MCMCAlgorithm,
    posterior::AbstractPosteriorDensity
)

Markov-chain instances, represented by objects of type MCMCIterator, are be created via

(spec::MCMCSpec)(chainid::Integer)

BAT.OnlineMvCov — Type

OnlineMvCov{T<:AbstractFloat,W} <: AbstractMatrix{T}

BAT-internal, not part of stable public API.

Implementation based on variance calculation Algorithms of Welford and West.

W must either be Weights (no bias correction) or one of AnalyticWeights, FrequencyWeights or ProbabilityWeights to specify the desired bias correction method.

BAT.OnlineMvMean — Type

OnlineMvMean{T<:AbstractFloat} <: AbstractVector{T}

BAT-internal, not part of stable public API.

Multivariate mean implemented via Kahan-Babuška-Neumaier summation.

BAT.OnlineUvMean — Type

OnlineUvMean{T<:AbstractFloat}

BAT-internal, not part of stable public API.

Univariate mean implemented via Kahan-Babuška-Neumaier summation.

BAT.OnlineUvVar — Type

OnlineUvVar{T<:AbstractFloat,W}

BAT-internal, not part of stable public API.

Implementation based on variance calculation Algorithms of Welford and West.

W must either be Weights (no bias correction) or one of AnalyticWeights, FrequencyWeights or ProbabilityWeights to specify the desired bias correction method.

BAT.PointCloud — Type

PointCloud{T<:AbstractFloat, I<:Integer}

AHMI-internal, not part of stable public API.

Stores the information of the points of an e.g. HyperRectVolume

Variables

'maxLogProb' : The maximum log. probability of one of the points inside the hyper-rectangle

'minLogProb' : The minimum log. probability of one of the points inside the hyper-rectangle

'maxWeightProb' : the weighted max. log. probability

'minWeightProb' : the weighted min. log. probability

'probfactor' : The probability factor of the hyper-rectangle

'probweightfactor' : The weighted probability factor

'points' : The number of points inside the hyper-rectangle

'pointIDs' : the IDs of the points inside the hyper-rectangle, might be empty because it is optional and costs performance

'searchres' : used to boost performance

BAT.SearchResult — Type

SearchResult{T<:AbstractFloat, I<:Integer}

AHMI-internal, not part of stable public API.

Stores the results of the space partitioning tree's search function

Variables

'pointIDs' : the IDs of samples found, might be empty because it is optional

'points' : The number of points found.

'maxLogProb' : the maximum log. probability of the points found.

'minLogProb' : the minimum log. probability of the points found.

'maxWeightProb' : the weighted minimum log. probability found.

'minWeightProb' : the weighted maximum log. probfactor found.

BAT.WhiteningResult — Type

WhiteningResult{T<:AbstractFloat}

AHMI-internal, not part of stable public API.

Stores the information obtained during the Whitening Process

Variables

'determinant' : The determinant of the whitening matrix

'targetprobfactor' : The suggested target probability factor

'whiteningmatrix' : The whitening matrix

'meanvalue' : the mean vector of the input data

BAT.apply_bounds — Function

apply_bounds(x::<:Real, lo::<:Real, hi::<:Real, boundary_type::BoundsType)

BAT-internal, not part of stable public API.

Apply lower/upper bound lo/hi to value x. boundary_type may be hard_bounds, cyclic_bounds or reflective_bounds.

apply_bounds(x::Real, interval::ClosedInterval, boundary_type::BoundsType)

BAT-internal, not part of stable public API.

Specify lower and upper bound via interval.

BAT.apply_bounds_and_eval_posterior_logval_strict! — Function

BAT.apply_bounds_and_eval_posterior_logval_strict!(
    posterior::AbstractPosteriorDensity,
    v::AbstractVector{<:Real}
)

BAT-internal, not part of stable public API.

First apply bounds to the parameters, then compute and return posterior log value. May modify v.

Guarantees that :

If parameters are still out of bounds after applying bounds, density_logval is not called for either prior or likelihood.

If density_logval for prior returns -Inf, density_logval is not called for likelihood.

In both cases, an exception is thrown.

BAT.apply_bounds_and_eval_posterior_logval! — Function

BAT.apply_bounds_and_eval_posterior_logval!(
    T::Type{<:Real},
    density::AbstractDensity,
    v::AbstractVector{<:Real}
)

BAT-internal, not part of stable public API.

First apply bounds to the parameters, then compute and return posterior log value. May modify v.

Guarantees that :

If parameters are still out of bounds after applying bounds, density_logval is not called for either prior or likelihood.

If density_logval for prior returns -Inf, density_logval is not called for likelihood.

In both cases, T(-Inf) is returned for both prior and posterior.

BAT.apply_bounds! — Function

apply_bounds!(x::AbstractVector, bounds::AbstractVarBounds)

BAT-internal, not part of stable public API.

Apply bounds to variate/parameters x.

BAT.autocrl — Function

autocrl(xv::AbstractVector{T}, kv::AbstractVector{Int} = Vector{Int}())

BAT-internal, not part of stable public API.

autocorrelation := Σ Cov[xi,x(i+k)]/Var[x]

Computes the autocorrelations at various leg k of the input vector (time series) xv. The vector kv is the collections of lags to take into account

BAT.bat_sampler — Function

bat_sampler(d::Distribution)

BAT-internal, not part of stable public API.

Tries to return a BAT-compatible sampler for Distribution d. A sampler is BAT-compatible if it supports random number generation using an arbitrary AbstractRNG:

rand(rng::AbstractRNG, s::SamplerType)
rand!(rng::AbstractRNG, s::SamplerType, x::AbstractArray)

If no specific method of bat_sampler is defined for the type of d, it will default to sampler(d), which may or may not return a BAT-compatible sampler.

BAT.bg_R_2sqr — Function

bg_R_2sqr(stats::AbstractVector{<:MCMCBasicStats}; corrected::Bool = false)

BAT-internal, not part of stable public API.

Brooks-Gelman $R_2^2$ for all DOF. If normality is assumed, 'corrected' should be set to true to account for the sampling variability.

BAT.calculate_localmode — Function

calculate_localmode(hist)

BAT-internal, not part of stable public API.

Calculates the modes of a 1d statsbase histogram. A vector containing the bin-center(s) of the heighest bin(s) is returned.

BAT.create_hypercube — Function

create_hypercube{T<:Real}(origin::Vector{T}, edgelength::T)::HyperRectVolume

AHMI-internal, not part of stable public API.

creates a hypercube shaped spatial volume

BAT.create_hyperrectangle — Function

create_hyperrectangle(...)

AHMI-internal, not part of stable public API.

This function creates a hyper-rectangle around each starting sample. It starts by building a hyper-cube and subsequently adapts each face individually, thus turning the hyper-cube into a hyper-rectangle. The faces are adjusted in a way to match the shape of the distribution as best as possible.

BAT.distribution_logpdf — Function

distribution_logpdf(
    p::AbstractArray,
    pdist::AbstractProposalDist,
    v_proposed::AbstractVector,
    v_current:::AbstractVector
)

BAT-internal, not part of stable public API.

Analog to distribution_logpdf!, but for a single variate/parameter vector.

BAT.distribution_logpdf! — Function

distribution_logpdf!(
    p::AbstractArray,
    pdist::AbstractProposalDist,
    v_proposed::Union{AbstractVector,VectorOfSimilarVectors},
    v_current:::Union{AbstractVector,VectorOfSimilarVectors}
)

BAT-internal, not part of stable public API.

Returns log(PDF) value of pdist for transitioning from current to proposed variate/parameter arguments values for a variate/parameter sets.

end

Input:

v_proposed: New values (column vectors)

v_current: Old values (column vectors)

Output is stored in

p: Array of PDF values, length must match, shape is ignored

Array size requirements:

size(v_current, 1) == size(v_proposed, 1) == length(pdist)

size(v_current, 2) == size(v_proposed, 2) or size(v_current, 2) == 1

size(v_proposed, 2) == length(p)

Implementations of distribution_logpdf! must be thread-safe.

BAT.drop_low_weight_samples — Function

drop_low_weight_samples(
    samples::DensitySampleVector,
    fraction::Real = 10^-4
)

BAT-internal, not part of stable public API.

Drop fraction of the total probability mass from samples to filter out the samples with the lowest weight.

BAT.effective_sample_size — Function

effective_sample_size(xv, w; ...)

BAT-internal, not part of stable public API.

Effective size estimation for a vector of samples xv. If a weight vector w is provided, the Kish approximation is applied.

By default computes the autocorrelation up to the square root of the number of entries in the vector, unless an explicit list of lags is provided (kv).

effective_sample_size(variates::AbstractArray, weights::AbstractVector; with_weights=true)

BAT-internal, not part of stable public API.

Effective size estimation for a (multidimensional) ElasticArray. By default applies the Kish approximation with the weigths available, but can be turned off (with_weights=false).

effective_sample_size(samples::DensitySampleVector; with_weights=true)

BAT-internal, not part of stable public API.

Effective size estimation for a (multidimensional) DensitySampleVector. By default applies the Kish approximation with the weigths available, but can be turned off (with_weights=false).

BAT.eval_density_logval — Function

eval_density_logval(
    density::AbstractDensity,
    v::AbstractVector{<:Real},
    shape::ValueShapes.AbstractValueShape
)

BAT-internal, not part of stable public API.

Evaluates density log-value via density_logval.

shape must be compatible with ValueShapes.varshape(density).

Checks that:

The variate shape of density (if known) matches shape of v.

The return value of density_logval is not NaN.

The return value of density_logval is less than +Inf.

BAT.find_hypercube_centers — Function

find_hypercube_centers(dataset::DataSet{T, I}, whiteningresult::WhiteningResult, settings::HMISettings)::Vector{I}

AHMI-internal, not part of stable public API.

finds possible starting points for the hyperrectangle creation

BAT.fromuhc — Function

fromuhc(X::AbstractVector, vol::SpatialVolume)
fromuhc(X::VectorOfSimilarVectors, vol::SpatialVolume)

BAT-internal, not part of stable public API.

Bijective transformation from unit hypercube to vol. See fromuhc!.

Use inv(fromuhc) to get the the inverse transformation.

BAT.fromuhc! — Function

fromuhc!(Y::AbstractVector, X::AbstractVector, vol::SpatialVolume)
fromuhc!(Y::VectorOfSimilarVectors, X::VectorOfSimilarVectors, vol::SpatialVolume)

BAT-internal, not part of stable public API.

Bijective transformation of coordinates X within the unit hypercube to coordinates Y in vol. If X and Y are matrices, the transformation is applied to the column vectors. Use Y === X to transform in-place.

Use inv(fromuhc!) to get the the inverse transformation.

BAT.fromui — Function

y = fromui(x::Real, lo::Real, hi::Real)
y = fromui(x::Real, lo_hi::ClosedInterval{<:Real})

BAT-internal, not part of stable public API.

Linear bijective transformation from the unit inverval (i.e. x ∈ 0..1) to y ∈ lo..hi.

Use inv(fromui) to get the the inverse transformation.

Use @inbounds to disable range checking on the input value.

BAT.getlikelihood — Function

getlikelihood(posterior::AbstractPosteriorDensity)::AbstractDensity

BAT-internal, not part of stable public API.

The likelihood density of posterior. The likelihood may or may not be normalized.

BAT.getprior — Function

getprior(posterior::AbstractPosteriorDensity)::AbstractDensity

BAT-internal, not part of stable public API.

The prior density of posterior. The prior may or may not be normalized.

BAT.gr_Rsqr — Function

gr_Rsqr(stats::AbstractVector{<:MCMCBasicStats})

BAT-internal, not part of stable public API.

Gelman-Rubin $\$R^2\$$ for all DOF.

BAT.hm_init — Function

hm_init!(result, settings)

AHMI-internal, not part of stable public API.

Sets the global multithreading setting and ensures that a minimum number of samples, dependent on the number of dimensions, are provided.

BAT.hm_integrate! — Function

hm_integrate!(result, settings = HMIPrecisionSettings())

AHMI-internal, not part of stable public API.

This function starts the adaptive harmonic mean integration. See arXiv:1808.08051 for more details. It needs a HMIData struct as input, which holds the samples, in form of a dataset, the integration volumes and other properties, required for the integration, and the final result.

BAT.hm_whiteningtransformation! — Function

hm_whiteningtransformation!(result, settings)

AHMI-internal, not part of stable public API.

Applies a whitening transformation to the samples. A custom whitening method can be used by overriding settings.whitening_function!

BAT.hyperrectangle_creationproccess! — Function

hyperrectangle_creationproccess!(...)

AHMI-internal, not part of stable public API.

This function assigns each thread its own hyper-rectangle to build, if in multithreading-mode.

BAT.integrate_hyperrectangle_cov — Function

Estimates reducedvolumehm quantity for one hyperrectangle including estimates from batches (for covariance calculations).

BAT.issymmetric_around_origin — Function

issymmetric_around_origin(d::Distribution)

BAT-internal, not part of stable public API.

Returns true (resp. false) if the Distribution is symmetric (resp. non-symmetric) around the origin.

BAT.log_volume — Function

log_volume(vol::SpatialVolume)

BAT-internal, not part of stable public API.

Get the logarithm of the volume of the space in vol.

BAT.mcmc_startval! — Function

BAT.mcmc_startval!(
    x::Union{AbstractVector{<:Real},VectorOfSimilarVectors{<:Real}},
    rng::AbstractRNG,
    posterior::AbstractPosteriorDensity,
    algorithm::MCMCAlgorithm
)::typeof(x)

BAT-internal, not part of stable public API.

Fill x a random initial argument suitable for posterior and algorithm. The default implementation will try to draw the initial argument value from the prior of the posterior.

BAT.modify_hypercube! — Function

create_hypercube!{T<:Real}(origin::Vector{T}, edgelength::T)::HyperRectVolume

AHMI-internal, not part of stable public API.

resizes a hypercube shaped spatial volume

BAT.modify_integrationvolume! — Function

modify_integrationvolume!(intvol::IntegrationVolume{T, I}, dataset::DataSet{T, I}, spvol::HyperRectVolume{T}, searchpts::Bool = true)

AHMI-internal, not part of stable public API.

updates an integration volume with new boundaries. Recalculates the pointcloud and volume.

BAT.proposal_rand! — Function

function proposal_rand!(
    rng::AbstractRNG,
    pdist::GenericProposalDist,
    v_proposed::Union{AbstractVector,VectorOfSimilarVectors},
    v_current::Union{AbstractVector,VectorOfSimilarVectors}
)

BAT-internal, not part of stable public API.

Generate one or multiple proposed variate/parameter vectors, based on one or multiple previous vectors.

Input:

rng: Random number generator to use

pdist: Proposal distribution to use

v_current: Old values (vector or column vectors, if a matrix)

Output is stored in

v_proposed: New values (vector or column vectors, if a matrix)

The caller must guarantee:

size(v_current, 1) == size(v_proposed, 1)

size(v_current, 2) == size(v_proposed, 2) or size(v_current, 2) == 1

v_proposed !== v_current (no aliasing)

Implementations of proposal_rand! must be thread-safe.

BAT.reduced_volume_hm — Function

reduced_volume_hm!(log_prob, sample_weights, volume_size, n_total, weight_total, bias_correction==true)

Estimate reduced volume harmonic mean for given arguments. Includes bias corection by default. Note: Weights are expected to be frequency weights (i.e. counts).

BAT.spatialvolume — Function

spatialvolume(b::VarVolumeBounds)::SpatialVolume

BAT-internal, not part of stable public API.

Returns the spatial volume that defines the variate/parameter bounds.

BAT.sum_first_dim — Function

@propagate_inbounds sum_first_dim(A::AbstractArray, j::Integer, ks::Integer...)

BAT-internal, not part of stable public API.

Calculate the equivalent of sum(A[:, j, ks...]).

@propagate_inbounds sum_first_dim(A::AbstractArray)

BAT-internal, not part of stable public API.

If A is a vector, return sum(A), else sum(A, 1)[:].

BAT.var_bounds — Function

var_bounds(density::DistLikeDensity)::AbstractVarBounds

BAT-internal, not part of stable public API.

Get the parameter bounds of density. Must not be missing.

BAT.wgt_effective_sample_size — Function

wgt_effective_sample_size(w::AbstractVector{T})

BAT-internal, not part of stable public API.

Kish's approximation for weighted samples effectivesamplesize estimation. Computes the weighting factor for weigthed samples, where w is the vector of weigths.