Internal API

Note

This is the documentation of BAT's internal API. The internal API is fully accessible to users, but all aspects of it are subject to change without deprecation. Functionalities of the internal API that, over time, turn out to require user access (e.g. to support advanced use cases) will be evolved to gain a stable interface and then promoted to the public API.

Types

BAT.AbstractProposalDist

BAT.AbstractSampleGenerator

BAT.BasicMvStatistics

BAT.DataSet

BAT.HMIData

BAT.HMISettings

BAT.IntegrationVolume

BAT.KDTreePartitioning

BAT.MCMCSampleGenerator

BAT.OnlineMvCov

BAT.OnlineMvMean

BAT.OnlineUvMean

BAT.OnlineUvVar

BAT.PointCloud

BAT.SearchResult

BAT.SpacePartTree

BAT.StandardMvNormal

BAT.StandardMvUniform

BAT.StandardUvNormal

BAT.StandardUvUniform

BAT.TransformedDensity

BAT.TruncatedDensity

BAT.VariateTransform

BAT.WhiteningResult

Functions and macros

BAT.argchoice_msg

BAT.bat_sampler

BAT.bg_R_2sqr

BAT.create_hypercube

BAT.create_hyperrectangle

BAT.density_logval_type

BAT.drop_low_weight_samples

BAT.eval_gradlogval

BAT.eval_logval

BAT.find_hypercube_centers

BAT.find_marginalmodes

BAT.fromuhc

BAT.fromuhc!

BAT.fromui

BAT.get_bin_centers

BAT.getlikelihood

BAT.getprior

BAT.gr_Rsqr

BAT.hm_init

BAT.hm_integrate!

BAT.hm_whiteningtransformation!

BAT.hyperrectangle_creationproccess!

BAT.integrate_hyperrectangle_cov

BAT.is_log_zero

BAT.issymmetric_around_origin

BAT.log_volume

BAT.log_zero_density

BAT.modify_hypercube!

BAT.modify_integrationvolume!

BAT.partition_space

BAT.proposal_rand!

BAT.proposaldist_logpdf

BAT.reduced_volume_hm

BAT.repetition_to_weights

BAT.spatialvolume

BAT.sum_first_dim

BAT.trunc_logpdf_ratio

BAT.truncate_density

BAT.truncate_dist_hard

BAT.var_bounds

Documentation

BAT.AbstractProposalDist — Type

abstract type AbstractProposalDist

BAT-internal, not part of stable public API.

The following functions must be implemented for subtypes:

BAT.proposaldist_logpdf

BAT.proposal_rand!

ValueShapes.totalndof, returning the number of DOF (i.e. dimensionality).

LinearAlgebra.issymmetric, indicating whether p(a -> b) == p(b -> a) holds true.

BAT.AbstractSampleGenerator — Type

abstract type AbstractSampleGenerator

BAT-internal, not part of stable public API.

Abstract super type for sample generators.

BAT.BasicMvStatistics — Type

BasicMvStatistics{T<:Real,W}

BAT-internal, not part of stable public API.

W must either be Weights (no bias correction) or one of AnalyticWeights, FrequencyWeights or ProbabilityWeights to specify the desired bias correction method.

BAT.DataSet — Type

DataSet{T<:AbstractFloat, I<:Integer}

AHMI-internal, not part of stable public API.

Holds the MCMC output. For construction use constructor: function DataSet{T<:Real}(data::Matrix{T}, logprob::Vector{T}, weights::Vector{T})

Variables

'data' : An P x N array with N data points with P parameters.

'logprob' : The logarithmic probability for each samples stored in an array

'weights' : How often each sample occurred. Set to an array of ones if working directly on MCMC output

'ids' : Array which is used to assign each sample to a batch, required for the cov. weighed uncertainty estimation

.sortids : an array of indices which stores the original ordering of the samples (the space partitioning tree reorders the samples), required to calculate an effective sample size.

'N' : number of samples

'P' : number of parameters

'nsubsets' : the number of batches

'iswhitened' : a boolean value which indicates whether the data set is iswhitened

'isnew' : a boolean value which indicates whether the data set was swapped out with a different one (it is possible to redo the integration with a different sample set using previously generated hyper-rectangles)

'partitioningtree' : The space partitioning tree, used to efficiently identify samples in a point cloud

'startingIDs' : The Hyper-Rectangle Seed Samples are stored in this array

'tolerance' : A threshold required for the hyper-rectangle creation process.

BAT.HMIData — Type

HMIData{T<:AbstractFloat, I<:Integer}

AHMI-internal, not part of stable public API.

Includes all the informations of the integration process, including a list of hyper-rectangles, the results of the whitening transformation, the starting ids, and the average number of points and volume of the created hyper-rectangles.

Variables

'dataset1' : Data Set 1

'dataset2' : Data Set 2

'whiteningresult' : contains the whitening matrix and its determinant, required to scale the final integral estimate

'volumelist1' : An array of integration volumes created using dataset1, but filled with samples from dataset2

'volumelist2' : An array of integration volumes created using dataset2, but filled with samples from dataset1

'cubelist1' : An array of small hyper-cubes created around seeding samples of dataset 1

'cubelist2' : An array of small hyper-cubes created around seeding samples of dataset 2

'iterations1' : The number of volume adapting iterations for the creating volumelist1

'iterations2' : The number of volume adapting iterations for the creating volumelist2

'rejectedrects1' : An array of ids, indicating which hyper-rectangles of volumelist1 were rejected due to trimming

'rejectedrects2' : An array of ids, indicating which hyper-rectangles of volumelist2 were rejected due to trimming

'integralestimates' : A dictionary containing the final integral estimates with uncertainty estimation using different uncertainty estimators. Also includes all intermediate results required for the integral estimate combination

BAT.HMISettings — Type

HMISettings

AHMI-internal, not part of stable public API.

holds the settings for the hm_integrate function. There are several default constructors available: HMIFastSettings() HMIStandardSettings() HMIPrecisionSettings()

#Variables

'whitening_method::Symbol' : which whitening method to use

'max_startingIDs::Integer' : influences how many starting ids are allowed to be generated

'maxstartingIDsfraction::AbstractFloat' : how many points are considered as possible starting points as a fraction of total points available

'rect_increase::AbstractFloat' : describes the procentual rectangle volume increase/decrease during hyperrectangle creation. Low values can increase the precision if enough points are available but can cause systematically wrong results if not enough points are available.

'useallrects::Bool' : All rectangles are used for the integration process no matter how big their overlap is. If enabled the rectangles are weighted by their overlap.

'useMultiThreading' : activate multithreading support.

'warning_minstartingids' : the required minimum amount of starting samples

'dotrimming' : determines whether the integral estimates are trimmed (1σ trim) before combining them into a final result (more robust)

'uncertaintyestimators' : A dictionary of different uncertainty estimator functions. Currently three functions are available: hmcombineresultslegacy! (outdated, overestimates uncertainty significantly in higher dimensions), hmcombineresultscovweighted! (very fast) and hmcombineresults_analyticestimation! (recommended)

end

BAT.IntegrationVolume — Type

IntegrationVolume{T<:AbstractFloat, I<:Integer}

AHMI-internal, not part of stable public API.

Variables

'pointcloud' : holds the point cloud of the integration volume

'spatialvolume' : the boundaries of the integration volume

'volume' : the volume

Hold the point cloud and the spatial volume for integration.

BAT.KDTreePartitioning — Type

struct KDTreePartitioning <: SpacePartitioningAlgorithm

BAT-internal, not part of stable public API.

K-D binary space partitioning algorithm. By default, all parameters are considered for partitioning. Partition parameters can be specified manually by using partition_dims argument. By default, bounds of the partitioning tree are extended to those given by prior. This can be changed by setting extend_bounds = false.

Constructors:

KDTreePartitioning(; fields...)

Fields:

partition_dims::Union{Array{Int64,1}, Symbol}

Default: :auto

extend_bounds::Bool

Default: true

BAT.MCMCSampleGenerator — Type

BAT.MCMCSampleGenerator

BAT-internal, not part of stable public API.

MCMC sample generator.

Constructors:

MCMCSampleGenerator(chain::AbstractVector{<:MCMCIterator})

BAT.OnlineMvCov — Type

OnlineMvCov{T<:AbstractFloat,W} <: AbstractMatrix{T}

BAT-internal, not part of stable public API.

Implementation based on variance calculation Algorithms of Welford and West.

W must either be Weights (no bias correction) or one of AnalyticWeights, FrequencyWeights or ProbabilityWeights to specify the desired bias correction method.

BAT.OnlineMvMean — Type

OnlineMvMean{T<:AbstractFloat} <: AbstractVector{T}

BAT-internal, not part of stable public API.

Multivariate mean implemented via Kahan-Babuška-Neumaier summation.

BAT.OnlineUvMean — Type

OnlineUvMean{T<:AbstractFloat}

BAT-internal, not part of stable public API.

Univariate mean implemented via Kahan-Babuška-Neumaier summation.

BAT.OnlineUvVar — Type

OnlineUvVar{T<:AbstractFloat,W}

BAT-internal, not part of stable public API.

Implementation based on variance calculation Algorithms of Welford and West.

W must either be Weights (no bias correction) or one of AnalyticWeights, FrequencyWeights or ProbabilityWeights to specify the desired bias correction method.

BAT.PointCloud — Type

PointCloud{T<:AbstractFloat, I<:Integer}

AHMI-internal, not part of stable public API.

Stores the information of the points of an e.g. HyperRectVolume

Variables

'maxLogProb' : The maximum log. probability of one of the points inside the hyper-rectangle

'minLogProb' : The minimum log. probability of one of the points inside the hyper-rectangle

'maxWeightProb' : the weighted max. log. probability

'minWeightProb' : the weighted min. log. probability

'probfactor' : The probability factor of the hyper-rectangle

'probweightfactor' : The weighted probability factor

'points' : The number of points inside the hyper-rectangle

'pointIDs' : the IDs of the points inside the hyper-rectangle, might be empty because it is optional and costs performance

'searchres' : used to boost performance

BAT.SearchResult — Type

SearchResult{T<:AbstractFloat, I<:Integer}

AHMI-internal, not part of stable public API.

Stores the results of the space partitioning tree's search function

Variables

'pointIDs' : the IDs of samples found, might be empty because it is optional

'points' : The number of points found.

'maxLogProb' : the maximum log. probability of the points found.

'minLogProb' : the minimum log. probability of the points found.

'maxWeightProb' : the weighted minimum log. probability found.

'minWeightProb' : the weighted maximum log. probfactor found.

BAT.SpacePartTree — Type

SpacePartTree

BAT-internal, not part of stable public API.

The structure stores a partitioning tree generated by any SpacePartitioningAlgorithm. Variables:

terminated_leaf : true if the tree node is terminal, false otherwise.

bounds : Low and high bound of the tree leaf.

left_child : The left child of the tree, missing is the node is terminal.

right_child : The right child of the tree, missing is the node is terminal.

cut_axis : Axis along which cut is performed, missing if the node is terminal.

cut_coordinate : Coordinate at which a cut is performed, missing if the node is terminal.

cost : The sum of the cost functions over leaves.

cost_part : The cost function of the current leaf, missing if the node is terminal..

BAT.StandardMvNormal — Type

StandardMvNormal{T<:Real} <: Distributions.AbstractMvNormal

A standard n-dimensional multivariate normal distribution with it's mean at the origin and an identity covariance matrix.

Constructor:

    StandardMvNormal(n::Integer)
    StandardMvNormal{T<:Real}(n::Integer)

BAT.StandardMvUniform — Type

StandardMvUniform{T<:Real} <: Distributions.Distribution{Multivariate,Continuous}

A standard n-dimensional multivariate uniform distribution, from zero to one in each dimension.

Constructor:

    StandardMvUniform(n::Integer)
    StandardMvUniform{T<:Real}(n::Integer)

BAT.StandardUvNormal — Type

StandardUvNormal{T<:Real} <: Distributions.Distribution{Univariate,Continuous}

A standard normal distribution with a mean of zero and a variance of one.

Constructor:

    StandardUvNormal()
    StandardUvNormal{T<:Real}()

BAT.StandardUvUniform — Type

StandardUvUniform{T<:Real} <: Distributions.Distribution{Univariate,Continuous}

A standard uniform distribution between zero and one.

Constructor:

    StandardUvUniform()
    StandardUvUniform{T<:Real}()

BAT.TransformedDensity — Type

TransformedDensity

BAT-internal, not part of stable public API.

BAT.TruncatedDensity — Type

TruncatedDensity

Constructor:

TruncatedDensity(D<:AbstractDensity, B<:VarVolumeBounds)

BAT-internal, not part of stable public API.

A density truncated to given bounds, forced to effectively zero outside of those bounds. In contrast to a truncated distribution, the density is not renormalized.

BAT.VariateTransform — Type

abstract type VariateTransform{VT<:AbstractValueShape,VF<:AbstractValueShape}

BAT-internal, not part of stable public API.

Abstract parameterized type for change-of-variables transformations.

Subtypes (e.g. SomeTrafo <: VariateTransform) must implement:

BAT.apply_vartrafo_impl(trafo::SomeTrafo, v)

BAT.apply_vartrafo_impl(inv_trafo::InverseVT{SomeTrafo}, v)

ValueShapes.varshape(trafo::SomeTrafo)

for real values and/or real-valued vectors v.

BAT.WhiteningResult — Type

WhiteningResult{T<:AbstractFloat}

AHMI-internal, not part of stable public API.

Stores the information obtained during the Whitening Process

Variables

'determinant' : The determinant of the whitening matrix

'targetprobfactor' : The suggested target probability factor

'whiteningmatrix' : The whitening matrix

'meanvalue' : the mean vector of the input data

BAT.argchoice_msg — Function

argchoice_msg(f::Base.Callable, argname::Val, x)

BAT-internal, not part of stable public API.

Generates an information message regarding the choice of value x for argument argname of function f.

The value x will often be the result of bat_default.

BAT.bat_sampler — Function

bat_sampler(d::Distribution)

BAT-internal, not part of stable public API.

Tries to return a BAT-compatible sampler for Distribution d. A sampler is BAT-compatible if it supports random number generation using an arbitrary AbstractRNG:

rand(rng::AbstractRNG, s::SamplerType)
rand!(rng::AbstractRNG, s::SamplerType, x::AbstractArray)

If no specific method of bat_sampler is defined for the type of d, it will default to sampler(d), which may or may not return a BAT-compatible sampler.

BAT.bg_R_2sqr — Function

bg_R_2sqr(stats::AbstractVector{<:MCMCBasicStats}; corrected::Bool = false)
bg_R_2sqr(samples::AbstractVector{<:DensitySampleVector}; corrected::Bool = false)

BAT-internal, not part of stable public API.

Brooks-Gelman R_2^2 for all DOF. If normality is assumed, 'corrected' should be set to true to account for the sampling variability.

BAT.create_hypercube — Function

create_hypercube{T<:Real}(origin::Vector{T}, edgelength::T)::HyperRectVolume

AHMI-internal, not part of stable public API.

creates a hypercube shaped spatial volume

BAT.create_hyperrectangle — Function

create_hyperrectangle(...)

AHMI-internal, not part of stable public API.

This function creates a hyper-rectangle around each starting sample. It starts by building a hyper-cube and subsequently adapts each face individually, thus turning the hyper-cube into a hyper-rectangle. The faces are adjusted in a way to match the shape of the distribution as best as possible.

BAT.density_logval_type — Function

BAT.density_logval_type(v::Any, density::AbstractDensity, T::Type{<:Real} = Float32)

BAT-internal, not part of stable public API.

Determine a suitable return type of log-density functions, given a variate v and an optional additional default result type T.

BAT.proposaldist_logpdf — Function

proposaldist_logpdf(
    p::AbstractArray,
    pdist::AbstractProposalDist,
    v_proposed::AbstractVector,
    v_current:::AbstractVector
)

BAT-internal, not part of stable public API.

Returns log(PDF) value of pdist for transitioning from current to proposed variate/parameters.

BAT.drop_low_weight_samples — Function

drop_low_weight_samples(
    samples::DensitySampleVector,
    fraction::Real = 10^-4
)

BAT-internal, not part of stable public API.

Drop fraction of the total probability mass from samples to filter out the samples with the lowest weight.

BAT.eval_gradlogval — Function

eval_gradlogval(density::AbstractDensity, v::AbstractVector{<:Real})
grad

BAT-internal, not part of stable public API.

Compute the log of the value of a multivariate density function, as well as the gradient of the log-value for the given variate/parameter-values.

Input:

density: density function

v: argument, i.e. variate / parameter-values

Returns a tuple of the log density value and it's gradient.

Note: This function should not be specialized for custom density types!

BAT.eval_logval — Function

eval_logval(
    density::AbstractDensity,
    v::Any,
    T::Type{:Real} = density_logval_type(v, density);
    use_bounds::Bool = true,
    strict::Bool = false
)::T

BAT-internal, not part of stable public API.

Evaluates density log-value via eval_logval_unchecked.

Throws an exception on any of these conditions:

The variate shape of density (if known) does not match the shape of v.

The return value of eval_logval_unchecked is NaN.

The return value of eval_logval_unchecked is an equivalent of positive infinity.

Options:

use_bounds: Return an equivalent of negative infinity if v is out of bounds. use_bounds should only be set to false when it is known that the v is within bounds.

strict: Throw an exception if v is out of bounds.

BAT.find_hypercube_centers — Function

find_hypercube_centers(dataset::DataSet{T, I}, whiteningresult::WhiteningResult, settings::HMISettings)::Vector{I}

AHMI-internal, not part of stable public API.

finds possible starting points for the hyperrectangle creation

BAT.find_marginalmodes — Function

find_marginalmodes(marg::MarginalDist)

BAT-internal, not part of stable public API.

Find the modes of a MarginalDist. Returns a vector of the bin-centers of the bin(s) with the heighest weight.

BAT.fromuhc — Function

fromuhc(X::AbstractVector, vol::SpatialVolume)
fromuhc(X::VectorOfSimilarVectors, vol::SpatialVolume)

BAT-internal, not part of stable public API.

Bijective transformation from unit hypercube to vol. See fromuhc!.

Use inv(fromuhc) to get the the inverse transformation.

BAT.fromuhc! — Function

fromuhc!(Y::AbstractVector, X::AbstractVector, vol::SpatialVolume)
fromuhc!(Y::VectorOfSimilarVectors, X::VectorOfSimilarVectors, vol::SpatialVolume)

BAT-internal, not part of stable public API.

Bijective transformation of coordinates X within the unit hypercube to coordinates Y in vol. If X and Y are matrices, the transformation is applied to the column vectors. Use Y === X to transform in-place.

Use inv(fromuhc!) to get the the inverse transformation.

BAT.fromui — Function

y = fromui(x::Real, lo::Real, hi::Real)
y = fromui(x::Real, lo_hi::ClosedInterval{<:Real})

BAT-internal, not part of stable public API.

Linear bijective transformation from the unit inverval (i.e. x ∈ 0..1) to y ∈ lo..hi.

Use inv(fromui) to get the the inverse transformation.

Use @inbounds to disable range checking on the input value.

BAT.get_bin_centers — Function

get_bin_centers(marg::MarginalDist)

BAT-internal, not part of stable public API.

Returns a vector of the bin-centers.

BAT.getlikelihood — Function

getlikelihood(posterior::AbstractPosteriorDensity)::AbstractDensity

BAT-internal, not part of stable public API.

The likelihood density of posterior. The likelihood may or may not be normalized.

BAT.getprior — Function

getprior(posterior::AbstractPosteriorDensity)::AbstractDensity

BAT-internal, not part of stable public API.

The prior density of posterior. The prior may or may not be normalized.

BAT.gr_Rsqr — Function

gr_Rsqr(stats::AbstractVector{<:MCMCBasicStats})
gr_Rsqr(samples::AbstractVector{<:DensitySampleVector})

BAT-internal, not part of stable public API.

Gelman-Rubin $$R^2$$ for all DOF.

BAT.hm_init — Function

hm_init!(result, settings)

AHMI-internal, not part of stable public API.

Sets the global multithreading setting and ensures that a minimum number of samples, dependent on the number of dimensions, are provided.

BAT.hm_integrate! — Function

hm_integrate!(result, settings = HMIPrecisionSettings())

AHMI-internal, not part of stable public API.

This function starts the adaptive harmonic mean integration. See arXiv:1808.08051 for more details. It needs a HMIData struct as input, which holds the samples, in form of a dataset, the integration volumes and other properties, required for the integration, and the final result.

BAT.hm_whiteningtransformation! — Function

hm_whiteningtransformation!(result, settings)

AHMI-internal, not part of stable public API.

Applies a whitening transformation to the samples. A custom whitening method can be used by overriding settings.whitening_function!

BAT.hyperrectangle_creationproccess! — Function

hyperrectangle_creationproccess!(...)

AHMI-internal, not part of stable public API.

This function assigns each thread its own hyper-rectangle to build, if in multithreading-mode.

BAT.integrate_hyperrectangle_cov — Function

Estimates reducedvolumehm quantity for one hyperrectangle including estimates from batches (for covariance calculations).

BAT.is_log_zero — Function

BAT.is_log_zero(x::Real, T::Type{<:Real} = typeof(x)}

BAT-internal, not part of stable public API.

Check if x is an equivalent of log of zero, resp. negative infinity, in respect to type T.

BAT.issymmetric_around_origin — Function

issymmetric_around_origin(d::Distribution)

BAT-internal, not part of stable public API.

Returns true (resp. false) if the Distribution is symmetric (resp. non-symmetric) around the origin.

BAT.log_volume — Function

log_volume(vol::SpatialVolume)

BAT-internal, not part of stable public API.

Get the logarithm of the volume of the space in vol.

BAT.log_zero_density — Function

BAT.log_zero_density(T::Type{<:Real})

log-density value to assume for regions of implicit zero density, e.g. outside of variate/parameter bounds/support.

Returns an equivalent of negative infinity.

BAT.modify_hypercube! — Function

create_hypercube!{T<:Real}(origin::Vector{T}, edgelength::T)::HyperRectVolume

AHMI-internal, not part of stable public API.

resizes a hypercube shaped spatial volume

BAT.modify_integrationvolume! — Function

modify_integrationvolume!(intvol::IntegrationVolume{T, I}, dataset::DataSet{T, I}, spvol::HyperRectVolume{T}, searchpts::Bool = true)

AHMI-internal, not part of stable public API.

updates an integration volume with new boundaries. Recalculates the pointcloud and volume.

BAT.partition_space — Function

partition_space(
	samples::DensitySampleVector,
	npartitions::Integer,
	algorithm::KDTreePartitioning
)

BAT-internal, not part of stable public API.

The function generates a space partition tree with the number of partitions given by npartitions, using KDTreePartitioning algorithm. Exploration samples are given by samples. The output contains SpacePartTree and the values of the cost function.

BAT.proposal_rand! — Function

function proposal_rand!(
    rng::AbstractRNG,
    pdist::GenericProposalDist,
    v_proposed::Union{AbstractVector,VectorOfSimilarVectors},
    v_current::Union{AbstractVector,VectorOfSimilarVectors}
)

BAT-internal, not part of stable public API.

Generate one or multiple proposed variate/parameter vectors, based on one or multiple previous vectors.

Input:

rng: Random number generator to use

pdist: Proposal distribution to use

v_current: Old values (vector or column vectors, if a matrix)

Output is stored in

v_proposed: New values (vector or column vectors, if a matrix)

The caller must guarantee:

size(v_current, 1) == size(v_proposed, 1)

size(v_current, 2) == size(v_proposed, 2) or size(v_current, 2) == 1

v_proposed !== v_current (no aliasing)

Implementations of proposal_rand! must be thread-safe.

BAT.reduced_volume_hm — Function

reduced_volume_hm!(log_prob, sample_weights, volume_size, n_total, weight_total, bias_correction==true)

Estimate reduced volume harmonic mean for given arguments. Includes bias corection by default. Note: Weights are expected to be frequency weights (i.e. counts).

BAT.repetition_to_weights — Function

repetition_to_weights(v::AbstractVector)

BAT-internal, not part of stable public API.

Drop (subsequently) repeated samples by adding weights.

BAT.spatialvolume — Function

spatialvolume(b::VarVolumeBounds)::SpatialVolume

BAT-internal, not part of stable public API.

Returns the spatial volume that defines the variate/parameter bounds.

BAT.sum_first_dim — Function

@propagate_inbounds sum_first_dim(A::AbstractArray, j::Integer, ks::Integer...)

BAT-internal, not part of stable public API.

Calculate the equivalent of sum(A[:, j, ks...]).

@propagate_inbounds sum_first_dim(A::AbstractArray)

BAT-internal, not part of stable public API.

If A is a vector, return sum(A), else sum(A, 1)[:].

BAT.trunc_logpdf_ratio — Function

BAT.trunc_logpdf_ratio(orig_dist::Distribution{TP}, trunc_dist::Distribution{TP})::AbstractFloat

BAT-internal, not part of stable public API.

Computes the log-ratio between the amplitude of the PDF of a truncated distribution and the original (untruncted) distribution, within the support of the truncated one.

The PDF of both distributions must have the same shape within the support of trunc_dist and may only differ in amplitude.

Mainly used to implement BAT.truncate_density, in conjunction with BAT.truncate_dist_hard. The result contributes to the logscalecorr factor of a [TruncatedDensity] that uses truncated distributions internally, to ensure the density does not get renormalized.

BAT.truncate_density — Function

BAT.truncate_density(density::AbstractDensity, bounds::AbstractVarBounds)::AbstractDensity

BAT-internal, not part of stable public API.

Truncate density to bounds, the resulting density will be effectively zero outside of those bounds. In contrast Distributions.truncated, truncate_density does not renormalize the density.

Currently implemented for BAT.HyperRectBounds only.

BAT.truncate_dist_hard — Function

BAT.truncate_dist_hard(dist::Distribution{Univariate}, bounds::Interval)::Distribution{Univariate}
BAT.truncate_dist_hard(dist::Distribution{Multivariate}, bounds::AbstractArray{<:Interval})::Distribution{Multivariate}

BAT-internal, not part of stable public API.

Generalized variant of Distributions.truncated - also handles multivariate distributions and operates on a best-effort basis: If distributions cannot be truncated, may return the original distribution.

Returns a NamedTuple

    (dist = trunc_dist, logscalecorr = logscalecorr)

with the truncated distribution and the log-PDF amplitude difference to the original (see BAT.trunc_logpdf_ratio).

Mainly used to implement BAT.truncate_density.

BAT.var_bounds — Function

var_bounds(density::DistLikeDensity)::AbstractVarBounds

BAT-internal, not part of stable public API.

Get the parameter bounds of density. Must not be missing.