API Documentation

This is the stable public API of BAT. Forward/backward compatibility follows Julia's semantic versioning rules.

Types

BAT.AHMIntegration

BAT.ARPWeighting

BAT.AbstractDensity

BAT.AbstractDensityTransformTarget

BAT.AbstractMCMCWeightingScheme

BAT.AbstractModeEstimator

BAT.AbstractPosteriorDensity

BAT.AbstractSamplingAlgorithm

BAT.AbstractTransformToInfinite

BAT.AbstractTransformToUnitspace

BAT.AbstractTransformedDensity

BAT.AbstractVariateTransform

BAT.AdaptiveMHTuning

BAT.AnyDensityLike

BAT.AnyIIDSampleable

BAT.AnySampleable

BAT.AutocorLenAlgorithm

BAT.BrooksGelmanConvergence

BAT.CholeskyPartialWhitening

BAT.CholeskyWhitening

BAT.DensityIdentityTransform

BAT.DensitySample

BAT.DensitySampleVector

BAT.DistLikeDensity

BAT.EffSampleSizeAlgorithm

BAT.EffSampleSizeFromAC

BAT.ExplicitInit

BAT.FullDensityTransform

BAT.GelmanRubinConvergence

BAT.GenericDensity

BAT.GeyerAutocorLen

BAT.HamiltonianMC

BAT.IIDSampling

BAT.InitFromIID

BAT.InitFromSamples

BAT.InitFromTarget

BAT.InitvalAlgorithm

BAT.IntegrationAlgorithm

BAT.LogDVal

BAT.MCMCAlgorithm

BAT.MCMCBurninAlgorithm

BAT.MCMCChainPoolInit

BAT.MCMCConvergenceTest

BAT.MCMCInitAlgorithm

BAT.MCMCIterator

BAT.MCMCMultiCycleBurnin

BAT.MCMCNoOpTuning

BAT.MCMCSampling

BAT.MCMCTuningAlgorithm

BAT.MHProposalDistTuning

BAT.MaxDensityLBFGS

BAT.MaxDensityNelderMead

BAT.MaxDensitySampleSearch

BAT.MetropolisHastings

BAT.ModeAsDefined

BAT.NoDensityTransform

BAT.NoWhitening

BAT.OrderedResampling

BAT.PosteriorDensity

BAT.PriorSubstitution

BAT.PriorToGaussian

BAT.PriorToUniform

BAT.RandResampling

BAT.RepetitionWeighting

BAT.SampledDensity

BAT.SokalAutocorLen

BAT.StatisticalWhitening

BAT.TransformAlgorithm

BAT.WhiteningAlgorithm

Functions and macros

BAT.bat_default

BAT.bat_eff_sample_size

BAT.bat_findmedian

BAT.bat_findmode

BAT.bat_initval

BAT.bat_integrate

BAT.bat_integrated_autocorr_len

BAT.bat_read

BAT.bat_sample

BAT.bat_transform

BAT.bat_write

BAT.eval_logval_unchecked

BAT.fft_autocor

BAT.fft_autocov

BAT.ladjof

BAT.logvalof

BAT.trafoof

ValueShapes.totalndof

ValueShapes.varshape

Documentation

BAT.logvalof — Function

logvalof(r::NamedTuple{(...,:log,...)})::Real
logvalof(r::LogDVal)::Real

Extract the log-density value from a result r.

Examples:

logvalof((..., log = log_value, ...)) == log_value
logvalof(LogDVal(log_value)) == log_value

BAT.bat_default — Function

bat_default(f::Base.Callable, argname::Symbol, objective...)
bat_default(f::Base.Callable, argname::Val, objective...)

Get the default value for argument argname of function f to use for objective(s).

objective(s) are mandatory arguments of function f that semantically constitute it's main objective(s), and that that a good default choice of optional arguments (e.g. choice of algorithm(s), etc.) may depend on. Which arguments are considered to be objectives is function-specific.

For example:

bat_default(bat_sample, :algorithm, density::PosteriorDensity) == MetropolisHastings()
bat_default(bat_sample, Val(:algorithm), samples::DensitySampleVector) == OrderedResampling()

BAT.bat_eff_sample_size — Function

bat_eff_sample_size(
    v::Union{AbstractVector{<:Real},AbstractVectorOfSimilarVectors{<:Real}},
    [algorithm::EffSampleSizeAlgorithm]
)

bat_eff_sample_size(
    smpls::DensitySampleVector,
    [algorithm::EffSampleSizeAlgorithm]
)

Estimate effective sample size estimation for variate series v, resp. density samples smpls, separately for each degree of freedom.

Returns a NamedTuple of the shape

(result = eff_sample_size, ...)

Result properties not listed here are algorithm-specific and are not part of the stable public API.

Note

Do not add add methods to bat_eff_sample_size, add methods to bat_eff_sample_size_impl instead.

BAT.bat_findmedian — Function

bat_findmedian(
    samples::DensitySampleVector
)

The function computes the median of marginalized samples.

Returns a NamedTuple of the shape

(result = v, ...)

Result properties not listed here are algorithm-specific and are not part of the stable public API.

Note

Do not add add methods to bat_findmedian, add methods to bat_findmedian_impl instead.

BAT.bat_findmode — Function

bat_findmode(
    target::BAT.AnySampleable,
    [algorithm::BAT.AbstractModeEstimator]
)::DensitySampleVector

Estimate the global mode of target.

Returns a NamedTuple of the shape

(result = X::DensitySampleVector, ...)

Result properties not listed here are algorithm-specific and are not part of the stable public API.

Note

Do not add add methods to bat_findmode, add methods to bat_findmode_impl instead.

BAT.bat_initval — Function

bat_initval(
    [rng::AbstractRNG,]
    target::BAT.AnyDensityLike,
    [algorithm::BAT.InitvalAlgorithm],
)::V

bat_initval(
    [rng::AbstractRNG,]
    target::BAT.AnyDensityLike,
    n::Integer,
    [algorithm::BAT.InitvalAlgorithm],
)::AbstractVector{<:V}

Generate one or n random initial/starting value(s) suitable for target.

Assuming the variates of target are of type T, returns a NamedTuple of the shape

(result = X::AbstractVector{T}, ...)

Result properties not listed here are algorithm-specific and are not part of the stable public API.

Note

Do not add add methods to bat_initval, add methods like

bat_initval_impl(rng::AbstractRNG, target::AnyDensityLike, algorithm::InitvalAlgorithm; kwargs...)
bat_initval_impl(rng::AbstractRNG, target::AnyDensityLike, n::Integer, algorithm::InitvalAlgorithm; kwargs...)

to bat_initval_impl instead.

BAT.bat_integrate — Function

bat_integrate(
    target::AnySampleable,
    [algorithm::IntegrationAlgorithm]
)::DensitySampleVector

Calculate the integral (evidence) of target.

Returns a NamedTuple of the shape

(result = X::Measurements.Measurement, ...)

Result properties not listed here are algorithm-specific and are not part of the stable public API.

Note

Do not add add methods to bat_integrate, add methods to bat_integrate_impl instead.

BAT.bat_integrated_autocorr_len — Function

bat_integrated_autocorr_len(
    v::Union{AbstractVector{<:Real},AbstractVectorOfSimilarVectors{<:Real}},
    algorithm::AutocorLenAlgorithm = GeyerAutocorLen()
)

Estimate the integrated autocorrelation length of variate series v, separately for each degree of freedom.

Returns a NamedTuple of the shape

(result = integrated_autocorr_len, ...)

Result properties not listed here are algorithm-specific and are not part of the stable public API.

Note

Do not add add methods to bat_integrated_autocorr_len, add methods to bat_integrated_autocorr_len_impl instead.

BAT.bat_read — Function

bat_read(filename::AbstractString, data)
bat_read(fn_with_subpath::Tuple{AbstractString, AbstractString}, data)

Read data from a file filename, resp. from an internal sub-path of the file (if supported by the file format), e.g. an HDF5 group.

Currently supported file formats are:

HDF5 with file extension ".h5" or ".hdf5"

Note

HDF5 I/O functionality is only available when the HDF5 package is loaded (e.g. via import HDF5).

BAT.bat_sample — Function

bat_sample(
    [rng::AbstractRNG],
    target::BAT.AnySampleable,
    [algorithm::BAT.AbstractSamplingAlgorithm]
)::DensitySampleVector

Draw samples from target using algorithm.

Depending on sampling algorithm, the samples may be independent or correlated (e.g. when using MCMC).

Returns a NamedTuple of the shape

(result = X::DensitySampleVector, ...)

Result properties not listed here are algorithm-specific and are not part of the stable public API.

Note

Do not add add methods to bat_sample, add methods to bat_sample_impl instead.

BAT.bat_write — Function

bat_write(filename::AbstractString, data)
bat_write(fn_with_subpath::Tuple{AbstractString, AbstractString}, data)

Write data to a file filename, resp. to an internal sub-path of the file (if supported by the file format), e.g. an HDF5 group.

Currently supported file formats are:

HDF5 with file extension ".h5" or ".hdf5"

Note

HDF5 I/O functionality is only available when the HDF5 package is loaded (e.g. via import HDF5).

BAT.ladjof — Function

ladjof(r::NamedTuple{(...,:ladj,...)})::Real

Extract the log(abs(det(jacobian))) value that is part of a result r.

Examples:

ladjof((..., ladj = some_ladj, ...)) == some_ladj
ladjof(trafo)(v) = trafo(v, )

BAT.trafoof — Function

trafoof(d::AbstractTransformedDensity)::AbstractDensity

Get the transform from parent(d) to d, so that

trafoof(d)(parent(d)) == d

BAT.fft_autocor — Function

fft_autocor(v::AbstractVector{<:Real})
fft_autocor(v::AbstractVectorOfSimilarVectors{<:Real})

Compute the autocorrelation function (ACF) of variate series v, separately for each degree of freedom.

Uses FFT, in contract to StatsBase.autocor.

BAT.fft_autocov — Function

fft_autocov(v::AbstractVector{<:Real})
fft_autocov(v::AbstractVectorOfSimilarVectors{<:Real})

Compute the autocovariance of of variate series v, separately for each degree of freedom.

Uses FFT, in contract to StatsBase.autocov.

BAT.eval_logval_unchecked — Function

BAT.eval_logval_unchecked(density::AbstractDensity, v::Any)

Compute log of the value of a multivariate density function for the given variate/parameter-values.

Input:

density: density function

v: argument, i.e. variate / parameter-values

Note: If eval_logval_unchecked is called with an argument that is out of bounds, the behavior is undefined. The result for arguments that are not within bounds is implicitly -Inf, but it is the caller's responsibility to handle these cases.

BAT.AbstractDensity — Type

abstract type AbstractDensity

Subtypes of AbstractDensity must implement the function

BAT.eval_logval_unchecked

For likelihood densities this is typically sufficient, since shape, and variate bounds will be inferred from the prior.

Densities with a known variate shape may also implement

ValueShapes.varshape

Densities with known variate bounds may also implement

BAT.var_bounds

Note

The function BAT.var_bounds is not part of the stable public BAT-API, it's name and arguments may change without deprecation.

BAT.AbstractDensityTransformTarget — Type

abstract type AbstractDensityTransformTarget

Abstract type for probability density transformation targets.

BAT.AbstractMCMCWeightingScheme — Type

abstract type AbstractMCMCWeightingScheme{T<:Real}

Abstract class for weighting schemes for MCMC samples.

Weight values will have type T.

BAT.AbstractPosteriorDensity — Type

abstract type AbstractPosteriorDensity <: AbstractDensity end

Abstract type for posterior probability densities.

BAT.AbstractTransformedDensity — Type

TransformedDensity <: AbstractDensity

Abstract type for transformed densities.

In addition to the AbstractDensity interface, subtypes must support the functions Base.parent and trafoof:

parent(d::SomeTransformedDensity)::AbstractDensity
trafoof(d::SomeTransformedDensity)::AbstractVariateTransform

BAT.AbstractTransformToInfinite — Type

abstract type AbstractTransformToInfinite <: AbstractDensityTransformTarget

Abstract type for density transformation targets that specify are transformation into unbounded space.

The density that results from such a transformation must be unbounded in all dimensions and that should taper off to zero somewhat smoothly in any direction.

BAT.AbstractTransformToUnitspace — Type

abstract type AbstractTransformToUnitspace <: AbstractDensityTransformTarget

Abstract type for density transformation targets that specify a transformation into the unit hypercube.

BAT.AbstractVariateTransform — Type

abstract type AbstractVariateTransform <: Function

Abstract type for change-of-variables transformations.

Subtypes (e.g. SomeTrafo <: AbstractVariateTransform) must support (with trafo::SomeTrafo):

    (trafo)(v_prev::SomeVariate) == v_new
    (trafo)(v_prev::SomeVariate, ladj_prev::Real)) == (v = v_new, ladj = ladj_new)
    (trafo)(s_prev::DensitySample)::DensitySample
    ((trafo2 ∘ trafo1)(v)::AbstractVariateTransform)(v) == trafo2(trafo1(v))
    inv(trafo)(trafo(v)) == v
    inv(inv(trafo)) == trafo

    ValueShapes.varshape(trafo)::ValueShapes.AbstractValueShape

with varshape(v_prev) == varshape(trafo).

ladj must be logabsdet(jacobian(trafo, v)).

BAT.AdaptiveMHTuning — Type

struct AdaptiveMHTuning <: MHProposalDistTuning

Adaptive MCMC tuning strategy for Metropolis-Hastings samplers.

Adapts the proposal function based on the acceptance ratio and covariance of the previous samples.

Constructors:

AdaptiveMHTuning(; fields...)

Fields:

λ::Float64

Controls the weight given to new covariance information in adapting the proposal distribution. Default: 0.5

α::IntervalSets.Interval{:closed,:closed,Float64}

Metropolis-Hastings acceptance ratio target, tuning will try to adapt the proposal distribution to bring the acceptance ratio inside this interval. Default: ClosedInterval(0.15, 0.35)

β::Float64

Controls how much the spread of the proposal distribution is widened/narrowed depending on the current MH acceptance ratio. Default: 1.5

c::IntervalSets.Interval{:closed,:closed,Float64}

Interval for allowed scale/spread of the proposal distribution. Default: ClosedInterval(0.0001, 100.0)

r::Real

Reweighting factor. Take accumulated sample statistics of previous tuning cycles into account with a relative weight of r. Set to 0 to completely reset sample statistics between each tuning cycle. Default: 0.5

BAT.AHMIntegration — Type

struct AHMIntegration <: IntegrationAlgorithm

Adaptive Harmonic Mean Integration algorithm (Caldwell et al.).

Constructors:

AHMIntegration(; fields...)

Fields:

whitening::WhiteningAlgorithm

Default: CholeskyPartialWhitening()

autocorlen::AutocorLenAlgorithm

Default: GeyerAutocorLen()

volumetype::Symbol

Default: :HyperRectangle

max_startingIDs::Int64

Default: 10000

max_startingIDs_fraction::Float64

Default: 2.5

rect_increase::Float64

Default: 0.1

warning_minstartingids::Int64

Default: 16

dotrimming::Bool

Default: true

uncertainty::Array{Symbol,1}

List of uncertainty estimation methods to use, first entry will be used for primary result. Valid values:

:cov: Integral uncertainty for integration regions is estimated based on covariance of integrals of subsets of samples in the regions

:ess: Integral uncertainty for integration regions is estimated based on estimated effective number of samples in each region.

Default: [:cov]

BAT.AnyDensityLike — Type

BAT.AnyDensityLike = Union{...}

Union of all types that BAT will accept as a probability density, resp. that convert(AbstractDenstiy, d) supports:

AbstractDensity

Distributions.Distribution

BAT.AnyIIDSampleable — Type

BAT.AnyIIDSampleable = Union{...}

Union of all distribution/density-like types that BAT can draw i.i.d. (independent and identically distributed) samples from:

DistLikeDensity

Distributions.Distribution

BAT.AnySampleable — Type

BAT.AnySampleable = Union{...}

Union of all types that BAT can sample from:

AbstractDensity

DensitySampleVector

Distributions.Distribution

BAT.ARPWeighting — Type

ARPWeighting{T<:AbstractFloat} <: AbstractMCMCWeightingScheme{T}

Sample weighting scheme suitable for accept/reject-based sampling algorithms (e.g. MetropolisHastings). Both accepted and rejected samples become part of the output, with a weight proportional to their original acceptance probability.

Constructors:

ARPWeighting()

BAT.AutocorLenAlgorithm — Type

abstract type AutocorLenAlgorithm

Abstract type for integrated autocorrelation length estimation algorithms.

BAT.bat_transform — Function

bat_transform(
    target::AbstractDensityTransformTarget,
    density::AnyDensityLike,
    [algorithm::TransformAlgorithm]
)::AbstractDensity

Transform density to another variate space defined/implied by target.

Returns a NamedTuple of the shape

(result = newdensity::AbstractDensity, trafo = vartrafo::AbstractVariateTransform, ...)

Result properties not listed here are algorithm-specific and are not part of the stable public API.

Note

Do not add add methods to bat_transform, add methods to bat_transform_impl instead.

BAT.BrooksGelmanConvergence — Type

struct BrooksGelmanConvergence <: MCMCConvergenceTest

Brooks-Gelman maximum R^2 convergence test.

Constructors:

BrooksGelmanConvergence(; fields...)

Fields:

threshold::Float64

Default: 1.1

corrected::Bool

Default: false

BAT.CholeskyPartialWhitening — Type

struct CholeskyPartialWhitening <: WhiteningAlgorithm

Whitening transformation based on a Cholesky transformation of the empirical sample covariance matrix.

Only transforms dimensions (degrees of freedom) for which the marginalized distribution asymptotically approaches zero in the positive and negative direction.

Constructors:

CholeskyPartialWhitening()

BAT.CholeskyWhitening — Type

struct CholeskyWhitening <: WhiteningAlgorithm

Whitening transformation based on a Cholesky transformation of the empirical sample covariance matrix.

Constructors:

CholeskyWhitening()

BAT.DensityIdentityTransform — Type

struct DensityIdentityTransform <: TransformAlgorithm

A no-op density transform algorithm that leaves any density unchanged.

Constructors:

DensityIdentityTransform()

BAT.DensitySample — Type

struct DensitySample

A weighted sample drawn according to an statistical density, e.g. a BAT.AbstractDensity.

Constructors:

DensitySampleVector(v::Any, logd::Real, weight::Real, info::Any, aux::Any)

Fields:

v::Any

variate value

logd::Real

log(density) value at v

weight::Real

Weight of the sample

info::Any

Additional info on the provenance of the sample. Content depends on the sampling algorithm.

aux::Any

Custom user-defined information attached to the sample.

Use DensitySampleVector to store vectors of multiple samples with an efficient column-based memory layout.

BAT.DensitySampleVector — Type

struct DensitySampleVector <: AbstractVector{<:DensitySample}

A vector of DensitySample elements.

DensitySampleVector is currently a type alias for StructArrays.StructArray{<:DensitySample,...}, though this is subject to change without deprecation.

Constructors:

    DensitySampleVector(
        (
            v::AbstractVector{<:AbstractVector{<:Real}},
            logd::AbstractVector{<:Real},
            weight::AbstractVector{<:Real},
            info::AbstractVector{<:Any},
            aux::AbstractVector{<:Any}
        )
    )

    DensitySampleVector(
            v::AbstractVector,
            logval::AbstractVector{<:Real};
            weight::Union{AbstractVector{<:Real}, Symbol} = fill(1, length(eachindex(v))),
            info::AbstractVector = fill(nothing, length(eachindex(v))),
            aux::AbstractVector = fill(nothing, length(eachindex(v)))
        )

With weight = :multiplicity repeated samples will be replaced by a single sample, with a weight equal to the number of repetitions.

BAT.DistLikeDensity — Type

abstract type DistLikeDensity <: AbstractDensity

A density that implements part of the Distributions.Distribution interface. Such densities are suitable for use as a priors.

Typically, custom priors should be implemented as subtypes of Distributions.Distribution. BAT will automatically wrap them in a subtype of DistLikeDensity.

Subtypes of DistLikeDensity are required to support more functionality than an AbstractDensity, but less than a Distribution{Multivariate,Continuous}.

A d::Distribution{Multivariate,Continuous} can be converted into (wrapped in) an DistLikeDensity via conv(DistLikeDensity, d).

The following functions must be implemented for subtypes:

BAT.eval_logval_unchecked

ValueShapes.varshape

Distributions.sampler

Statistics.cov

BAT.var_bounds

Note

The function BAT.var_bounds is not part of the stable public BAT-API and subject to change without deprecation.

BAT.EffSampleSizeAlgorithm — Type

abstract type EffSampleSizeAlgorithm

Abstract type for integrated autocorrelation length estimation algorithms.

BAT.EffSampleSizeFromAC — Type

struct EffSampleSizeFromAC <: EffSampleSizeAlgorithm

Effective sample size estimation based on the integrated autocorrelation length of the samples.

Constructors:

EffSampleSizeFromAC(; fields...)

Fields:

acalg::AutocorLenAlgorithm

Default: GeyerAutocorLen()

BAT.ExplicitInit — Type

struct ExplicitInit <: InitvalAlgorithm

Uses initial values from a given vector of one or more values/variates. The values are used in the order they appear in the vector, not randomly.

Constructors:

ExplicitInit(; fields...)

Fields:

xs::AbstractArray{T,1} where T

BAT.FullDensityTransform — Type

struct FullDensityTransform <: TransformAlgorithm

Transform the density as a whole a given specified target space. Operations that use the gradient of the density will require to the log(abs(jacobian)) of the transformation to be auto-differentiable.

Constructors:

FullDensityTransform()

BAT.GelmanRubinConvergence — Type

struct GelmanRubinConvergence <: MCMCConvergenceTest

Gelman-Rubin maximum R^2 convergence test.

Constructors:

GelmanRubinConvergence(; fields...)

Fields:

threshold::Float64

Default: 1.1

BAT.GeyerAutocorLen — Type

struct GeyerAutocorLen <: AutocorLenAlgorithm

Integrated autocorrelation length estimation based on Geyer’s initial monotone sequence criterion

See C. J. Geyer, "Practical Markov Chain Monte Carlo" (1992) and C. J. Geyer, "Introduction to Markov Chain Monte Carlo" (2011).

Constructors:

GeyerAutocorLen()

The same algorithm is used by STAN (v2.21) and MCMCChains.jl (v3.0, function ess_rhat).

BAT.HamiltonianMC — Type

struct HamiltonianMC <: MCMCAlgorithm

Hamiltonian Monte Carlo sampling algorithm.

The gradient of the target density is computed via auto-differentiation.

Note: The fields of HamiltonianMC are still subject to change, and not

yet part of stable public API!*

Constructors:

HamiltonianMC(; fields...)

Fields:

metric::BAT.HMCMetric

Default: DiagEuclideanMetric()

gradient::Module

Default: ForwardDiff

integrator::BAT.HMCIntegrator

Default: LeapfrogIntegrator()

proposal::BAT.HMCProposal

Default: NUTS()

adaptor::BAT.HMCAdaptor

Default: StanHMCAdaptor()

BAT.IIDSampling — Type

struct IIDSampling <: AbstractSamplingAlgorithm

Sample via Random.rand. Only supported for posteriors of type Distributions.MultivariateDistribution and BAT.DistLikeDensity.

Constructors:

IIDSampling(; fields...)

Fields:

nsamples::Int64

Default: 10 ^ 5

BAT.InitFromIID — Type

struct InitFromIID <: InitvalAlgorithm

Generates initial values for sampling, optimization, etc. by random resampling from a given set of samples.

Constructors:

InitFromIID()

BAT.InitFromSamples — Type

struct InitFromSamples <: InitvalAlgorithm

Generates initial values for sampling, optimization, etc. by direct sampling from a given i.i.d. sampleable source.

Constructors:

InitFromSamples()

BAT.InitFromTarget — Type

struct InitFromTarget <: InitvalAlgorithm

Generates initial values for sampling, optimization, etc. by direct i.i.d. sampling a suitable component of that target density (e.g. it's prior) that supports it.

If the target is supports direct i.i.d. sampling, e.g. because it is a distribution, initial values are sampled directly from the target.

If the target is a posterior density, initial values are sampled from the prior (or the prior's prior if the prior is a posterior itself, etc.).

If the target is a sampled density, initial values are (re-)sampled from the available samples.

Constructors:

InitFromTarget()

BAT.InitvalAlgorithm — Type

abstract type BAT.InitvalAlgorithm

Abstract type for BAT initial/starting value generation algorithms.

Many algorithms in BAT, like MCMC and optimization, need initial/starting values.

BAT.IntegrationAlgorithm — Type

abstract type IntegrationAlgorithm

Abstract type for integration algorithms.

BAT.LogDVal — Type

LogDVal{T<:Real}

Represent the logarithm of the value of a statistical density at some point. LogDVal provides means to unambiguously distinguish between linear and logarithmic values of a density.

Constructor:

LogDVal(logd::Real)

Use logvalof to extract the actual log-density value from a LogDVal:

    logvalof(LogDVal(d)) == d

BAT.MaxDensityLBFGS — Type

struct MaxDensityLBFGS <: AbstractModeEstimator

Estimates the mode of a probability density using LBFGS optimization (currently via Optim.jl, subject to change).

The gradient of the target density is computed via auto-differentiation.

Constructors:

MaxDensityLBFGS(; fields...)

Fields:

trafo::AbstractDensityTransformTarget

Default: PriorToGaussian()

init::InitvalAlgorithm

Default: InitFromTarget()

BAT.MaxDensityNelderMead — Type

struct MaxDensityNelderMead <: AbstractModeEstimator

Estimates the mode of a probability density using Nelder-Mead optimization (currently via Optim.jl, subject to change).

Constructors:

MaxDensityNelderMead(; fields...)

Fields:

trafo::AbstractDensityTransformTarget

Default: NoDensityTransform()

init::InitvalAlgorithm

Default: InitFromTarget()

BAT.MaxDensitySampleSearch — Type

MaxDensitySampleSearch <: AbstractModeEstimator

Constructors:

MaxDensitySampleSearch()

Estimate the mode as the variate with the highest posterior density value within a given set of samples.

BAT.MCMCAlgorithm — Type

abstract type MCMCAlgorithm

Abstract type for Markov chain Monte Carlo algorithms.

To implement a new MCMC algorithm, subtypes of both MCMCAlgorithm and MCMCIterator are required.

Note

The details of the MCMCIterator and MCMCAlgorithm API required to implement a new MCMC algorithm currently do not (yet) form part of the stable API and are subject to change without deprecation.

BAT.MCMCBurninAlgorithm — Type

abstract type MCMCBurninAlgorithm

Abstract type for MCMC burn-in algorithms.

BAT.MCMCChainPoolInit — Type

struct MCMCChainPoolInit <: MCMCInitAlgorithm

MCMC chain pool initialization strategy.

Constructors:

MCMCChainPoolInit(; fields...)

Fields:

init_tries_per_chain::IntervalSets.Interval{:closed,:closed,Int64}

Default: ClosedInterval(8, 128)

nsteps_init::Int64

Default: 1000

BAT.MCMCConvergenceTest — Type

abstract type MCMCConvergenceTest

Abstract type for MCMC convergence testing algorithms.

BAT.MCMCInitAlgorithm — Type

abstract type MCMCInitAlgorithm

Abstract type for MCMC initialization algorithms.

BAT.MCMCIterator — Type

abstract type MCMCIterator end

Represents the current state of an MCMC chain.

Note

The details of the MCMCIterator and MCMCAlgorithm API (see below) currently do not form part of the stable API and are subject to change without deprecation.

To implement a new MCMC algorithm, subtypes of both MCMCAlgorithm and MCMCIterator are required.

The following methods must be defined for subtypes of MCMCIterator (e.g. SomeMCMCIter<:MCMCIterator):

BAT.getalgorithm(chain::SomeMCMCIter)::MCMCAlgorithm

BAT.getdensity(chain::SomeMCMCIter)::AbstractDensity

BAT.getrng(chain::SomeMCMCIter)::AbstractRNG

BAT.mcmc_info(chain::SomeMCMCIter)::MCMCIteratorInfo

BAT.nsteps(chain::SomeMCMCIter)::Int

BAT.nsamples(chain::SomeMCMCIter)::Int

BAT.current_sample(chain::SomeMCMCIter)::DensitySample

BAT.sample_type(chain::SomeMCMCIter)::Type{<:DensitySample}

BAT.samples_available(chain::SomeMCMCIter, nonzero_weights::Bool = false)::Bool

BAT.get_samples!(samples::DensitySampleVector, chain::SomeMCMCIter, nonzero_weights::Bool)::typeof(samples)

BAT.next_cycle!(chain::SomeMCMCIter)::SomeMCMCIter

BAT.mcmc_step!(
    chain::SomeMCMCIter
    callback::Function,
)::nothing

The following methods are implemented by default:

getalgorithm(chain::MCMCIterator)
getdensity(chain::MCMCIterator)
DensitySampleVector(chain::MCMCIterator)
mcmc_iterate!(chain::MCMCIterator, ...)
mcmc_iterate!(chains::AbstractVector{<:MCMCIterator}, ...)
isvalidchain(chain::MCMCIterator)
isviablechain(chain::MCMCIterator)

BAT.MCMCMultiCycleBurnin — Type

struct MCMCMultiCycleBurnin <: MCMCBurninAlgorithm

A multi-cycle MCMC burn-in algorithm.

Constructors:

MCMCMultiCycleBurnin(; fields...)

Fields:

nsteps_per_cycle::Int64

Default: 10000

max_ncycles::Int64

Default: 30

BAT.MCMCNoOpTuning — Type

MCMCNoOpTuning <: MCMCTuningAlgorithm

No-op tuning, marks MCMC chains as tuned without performing any other changes on them. Useful if chains are pre-tuned or tuning is an internal part of the MCMC sampler implementation.

BAT.MCMCSampling — Type

struct MCMCSampling <: AbstractSamplingAlgorithm

Samples a probability density using Markov chain Monte Carlo.

Constructors:

MCMCSampling(; fields...)

Fields:

mcalg::MCMCAlgorithm

Default: MetropolisHastings()

trafo::AbstractDensityTransformTarget

Default: bat_default(MCMCSampling, Val(:trafo), mcalg)

nchains::Int64

Default: 4

nsteps::Int64

Default: bat_default(MCMCSampling, Val(:nsteps), mcalg, trafo, nchains)

init::MCMCInitAlgorithm

Default: MCMCChainPoolInit(nsteps_init = max(div(nsteps, 100), 250))

burnin::MCMCBurninAlgorithm

Default: MCMCMultiCycleBurnin(nstepspercycle = max(div(nsteps, 10), 2500))

convergence::MCMCConvergenceTest

Default: BrooksGelmanConvergence()

strict::Bool

Default: true

store_burnin::Bool

Default: false

nonzero_weights::Bool

Default: true

callback::Function

Default: nop_func

BAT.MCMCTuningAlgorithm — Type

abstract type MCMCTuningAlgorithm

Abstract type for MCMC tuning algorithms.

BAT.MetropolisHastings — Type

struct MetropolisHastings <: MCMCAlgorithm

Metropolis-Hastings MCMC sampling algorithm.

Constructors:

MetropolisHastings(; fields...)

Fields:

proposal::BAT.ProposalDistSpec

Default: MvTDistProposal()

weighting::AbstractMCMCWeightingScheme

Default: RepetitionWeighting()

tuning::MHProposalDistTuning

Default: AdaptiveMHTuning()

BAT.MHProposalDistTuning — Type

abstract type MHProposalDistTuning

Abstract type for Metropolis-Hastings tuning strategies for proposal distributions.

BAT.ModeAsDefined — Type

struct ModeAsDefined <: AbstractModeEstimator

Get the mode as defined by the density, resp. the underlying distribution (if available), via StatsBase.mode.

Constructors:

ModeAsDefined()

BAT.NoDensityTransform — Type

struct NoDensityTransform <: AbstractDensityTransformTarget

The identity density transformation target, specifies that densities should not be transformed.

Constructors:

NoDensityTransform()

BAT.NoWhitening — Type

struct NoWhitening <: WhiteningAlgorithm

No-op whitening transformation, leaves samples unchanged.

Constructors:

NoWhitening()

BAT.OrderedResampling — Type

struct OrderedResampling <: AbstractSamplingAlgorithm

Efficiently resamples from a given series of samples, keeping the order of samples.

Can be used to efficiently convert weighted samples into samples with unity weights.

Constructors:

OrderedResampling(; fields...)

Fields:

nsamples::Int64

Default: 10 ^ 5

BAT.PosteriorDensity — Type

struct PosteriorDensity{
    Li<:AbstractDensity,
    Pr<:DistLikeDensity,
    ...
} <: AbstractPosteriorDensity

A representation of a PosteriorDensity, based a likelihood and prior. Likelihood and prior be accessed via

getlikelihood(posterior::PosteriorDensity)::Li
getprior(posterior::PosteriorDensity)::Pr

Constructors:

PosteriorDensity(likelihood, prior)

likelihood and prior must be convertible to an AbstractDensity.

Fields:

likelihood::AbstractDensity

prior::AbstractDensity

parbounds::BAT.AbstractVarBounds

parshapes::AbstractValueShape

Note

Fields parbounds and parbounds do not form part of the stable public API and are subject to change without deprecation.

BAT.PriorSubstitution — Type

struct PriorSubstitution <: TransformAlgorithm

Substitute the prior by a given distribution and transform the likelihood accordingly. The log(abs(jacobian)) of the transformation does not need to be auto-differentiable even for operations that use the gradient of the posterior.

Constructors:

PriorSubstitution()

BAT.PriorToGaussian — Type

struct PriorToGaussian <: AbstractTransformToInfinite

Specifies that posterior densities should be transformed in a way that makes their pior equivalent to a standard multivariate normal distribution with an identity covariance matrix.

Constructors:

PriorToGaussian()

BAT.PriorToUniform — Type

struct PriorToUniform <: AbstractTransformToUnitspace

Specifies that posterior densities should be transformed in a way that makes their pior equivalent to a uniform distribution over the unit hypercube.

Constructors:

PriorToUniform()

BAT.RandResampling — Type

struct RandResampling <: AbstractSamplingAlgorithm

Resamples from a given set of samples.

Constructors:

RandResampling(; fields...)

Fields:

nsamples::Int64

Default: 10 ^ 5

BAT.RepetitionWeighting — Type

struct RepetitionWeighting{T<:AbstractFloat} <: AbstractMCMCWeightingScheme{T}

Sample weighting scheme suitable for sampling algorithms which may repeated samples multiple times in direct succession (e.g. MetropolisHastings). The repeated sample is stored only once, with a weight equal to the number of times it has been repeated (e.g. because a Markov chain has not moved during a sampling step).

Constructors:

RepetitionWeighting()

BAT.SampledDensity — Type

struct SampledDensity

Stores a density and samples drawn from it.

A report on the density, samples, properties of variates/parameters, etc. can be generated via Base.show.

Constructors:

SampledDensity(density::AbstractPosteriorDensity, samples::DensitySampleVector)

Fields:

density::AbstractDensity

samples::DensitySamplesVector

_stats::BAT.MCMCBasicStats

_generator::AbstractSampleGenerator

Note

Fields _stats and _generator do not form part of the stable public API and are subject to change without deprecation.

This type is likely to evolve into a subtype of AbstractDensity in future versions.

BAT.SokalAutocorLen — Type

struct SokalAutocorLen <: AutocorLenAlgorithm

Integrated autocorrelation length estimation based on the automated windowing procedure descibed in A. D. Sokal, "Monte Carlo Methods in Statistical Mechanics" (1996)

Same procedure is used by the emcee Python package (v3.0).

Constructors:

SokalAutocorLen(; fields...)

Fields:

c::Int64

Step size for window search Default: 5

BAT.StatisticalWhitening — Type

struct StatisticalWhitening <: WhiteningAlgorithm

Whitening transformation based statistical whitening. CholeskyPartialWhitening

Constructors:

StatisticalWhitening()

BAT.TransformAlgorithm — Type

abstract type TransformAlgorithm

Abstract type for density transformation algorithms.

BAT.WhiteningAlgorithm — Type

abstract type WhiteningAlgorithm

Abstract type for sample whitening algorithms.

BAT.AbstractModeEstimator — Type

abstract type BAT.AbstractModeEstimator

Abstract type for BAT optimization algorithms.

A typical application for optimization in BAT is mode estimation (see bat_findmode),

BAT.AbstractSamplingAlgorithm — Type

abstract type BAT.AbstractSamplingAlgorithm

Abstract type for BAT sampling algorithms. See bat_sample.

BAT.GenericDensity — Type

GenericDensity{F<:Function} <: AbstractDensity

Constructors:

GenericDensity(f)

Turns the density function f into a BAT-compatible AbstractDensity. The return type of f(v) must supported by logvalof.

It must be safe to execute f in parallel on multiple threads and processes.

ValueShapes.totalndof — Function

ValueShapes.totalndof(density::AbstractDensity)::Union{Int,Missing}

Get the number of degrees of freedom of the variates of density. May return missing, if the shape of the variates is not fixed.

ValueShapes.totalndof(density::DistLikeDensity)::Int

Get the number of degrees of freedom of the variates of density. Must not be missing, a DistLikeDensity must have a fixed variate shape.

ValueShapes.varshape — Function

ValueShapes.varshape(
    density::AbstractDensity
)::Union{ValueShapes.AbstractValueShape,Missing}

ValueShapes.varshape(
    density::DistLikeDensity
)::ValueShapes.AbstractValueShape

Get the shapes of the variates of density.

For prior densities, the result must not be missing, but may be nothing if the prior only supports unshaped variate/parameter vectors.