API Documentation

This is the stable public API of BAT. Forward/backward compatibility follows Julia's semantic versioning rules.

Types

BAT.AHMC

BAT.AHMIntegration

BAT.ARPWeighting

BAT.AbstractDensity

BAT.AbstractDensityValue

BAT.AbstractMCMCTuningStrategy

BAT.AbstractModeEstimator

BAT.AbstractPosteriorDensity

BAT.AbstractSamplingAlgorithm

BAT.AbstractWeightingScheme

BAT.AdaptiveMetropolisTuning

BAT.AutocorLenAlgorithm

BAT.BrooksGelmanConvergence

BAT.CholeskyPartialWhitening

BAT.CholeskyWhitening

BAT.DensitySample

BAT.DensitySampleVector

BAT.DistLikeDensity

BAT.GelmanRubinConvergence

BAT.GenericDensity

BAT.GeyerAutocorLen

BAT.IIDSampling

BAT.IntegrationAlgorithm

BAT.LinDVal

BAT.LogDVal

BAT.MCMCAlgorithm

BAT.MCMCBurninStrategy

BAT.MCMCInitStrategy

BAT.MCMCIterator

BAT.MaxDensityLBFGS

BAT.MaxDensityNelderMead

BAT.MaxDensitySampleSearch

BAT.MetropolisHastings

BAT.ModeAsDefined

BAT.NoWhitening

BAT.OrderedResampling

BAT.PosteriorDensity

BAT.RandResampling

BAT.RepetitionWeighting

BAT.SokalAutocorLen

BAT.StatisticalWhitening

BAT.WhiteningAlgorithm

Functions and macros

BAT.bat_default

BAT.bat_eff_sample_size

BAT.bat_findmedian

BAT.bat_findmode

BAT.bat_integrate

BAT.bat_integrated_autocorr_len

BAT.bat_marginalmode

BAT.bat_read

BAT.bat_rng

BAT.bat_sample

BAT.bat_write

BAT.fft_autocor

BAT.fft_autocov

BAT.logvalgradof

BAT.logvalof

BAT.logvalof_unchecked

ValueShapes.totalndof

ValueShapes.varshape

Documentation

BAT.bat_default — Function

bat_default(f::Function, argname::Symbol, objective...)
bat_default(f::Function, argname::Val, objective...)

Get the default value for argument argname of function f to use for objective(s).

objective(s) are mandatory arguments of function f that semantically constitute it's main objective(s), and that that a good default choice of optional arguments (e.g. choice of algorithm(s), etc.) may depend on. Which arguments are considered to be objectives is function-specific.

For example:

bat_default(bat_sample, :algorithm, density::PosteriorDensity) == MetropolisHastings()
bat_default(bat_sample, Val(:algorithm), samples::DensitySampleVector) == OrderedResampling()

BAT.bat_eff_sample_size — Function

bat_eff_sample_size(
    v::Union{AbstractVector{<:Real},AbstractVectorOfSimilarVectors{<:Real}},
    [algorithm::AutocorLenAlgorithm]
)

bat_eff_sample_size(
    smpls::DensitySampleVector,
    [algorithm::AutocorLenAlgorithm]
)

Estimate effective sample size estimation for variate series v, resp. density samples smpls, separately for each degree of freedom.

Returns a NamedTuple of the shape

(result = X::AbstractVector{<:Real}, ...)

Result properties not listed here are algorithm-specific and are not part of the stable BAT API.

Note

Do not add add methods to bat_eff_sample_size, add methods to bat_eff_sample_size_impl instead (same arguments).

BAT.bat_findmedian — Function

bat_findmedian(
    samples::DensitySampleVector
)

The function computes the median of marginalized samples.

Returns a NamedTuple of the shape

(result = v, ...)

Result properties not listed here are algorithm-specific and are not part of the stable BAT API.

Note

Do not add add methods to bat_findmedian, add methods to bat_findmedian_impl instead (same arguments).

BAT.bat_findmode — Function

bat_findmode(
    target::BAT.AnySampleable,
    [algorithm::BAT.AbstractModeEstimator];
    initial_mode::Union{Missing,DensitySampleVector,Any} = missing
)::DensitySampleVector

Estimate the global mode of target.

Returns a NamedTuple of the shape

(result = X::DensitySampleVector, ...)

Result properties not listed here are algorithm-specific and are not part of the stable BAT API.

Note

Do not add add methods to bat_findmode, add methods to bat_findmode_impl instead (same arguments).

BAT.bat_integrate — Function

bat_integrate(
    target::AnySampleable,
    [algorithm::IntegrationAlgorithm]
)::DensitySampleVector

Calculate the integral (evidence) of target.

Returns a NamedTuple: (result = x::Measurement.Measurement, ...)

Result properties not listed here are algorithm-specific and are not part of the stable BAT API.

Returns a NamedTuple of the shape

(result = X::AbstractVector{<:Real}, ...)

Result properties not listed here are algorithm-specific and are not part of the stable BAT API.

Note

Do not add add methods to bat_integrate, add methods to bat_integrate_impl instead (same arguments).

BAT.bat_integrated_autocorr_len — Function

bat_integrated_autocorr_len(
    v::Union{AbstractVector{<:Real},AbstractVectorOfSimilarVectors{<:Real}},
    algorithm::AutocorLenAlgorithm = GeyerAutocorLen()
)

Estimate the integrated autocorrelation length of variate series v, separately for each degree of freedom.

Returns a NamedTuple of the shape

(result = integrated_autocorr_len, ...)

Result properties not listed here are algorithm-specific and are not part of the stable BAT API.

Note

Do not add add methods to bat_integrated_autocorr_len, add methods to bat_integrated_autocorr_len_impl instead (same arguments).

BAT.bat_marginalmode — Function

bat_marginalmode(
    samples::DensitySampleVector;
    nbins::Union{Integer, Symbol} = 200
)::DensitySampleVector

Estimates a local mode of samples by finding the maximum of marginalized posterior for each dimension.

Returns a NamedTuple of the shape

(result = X::DensitySampleVector, ...)

nbins specifies the number of bins that are used for marginalization. The default value is nbins=200. The optimal number of bins can be estimated using the following keywords:

:sqrt — Square-root choice

:sturges — Sturges' formula

:rice — Rice Rule

:scott — Scott's normal reference rule

:fd — Freedman–Diaconis rule

Returns a NamedTuple of the shape

(result = X::DensitySampleVector, ...)

Result properties not listed here are algorithm-specific and are not part of the stable BAT API.

Note

Do not add add methods to bat_marginalmode, add methods to bat_marginalmode_impl instead (same arguments).

BAT.bat_read — Function

bat_read(filename::AbstractString, data)
bat_read(fn_with_subpath::Tuple{AbstractString, AbstractString}, data)

Read data from a file filename, resp. from an internal sub-path of the file (if supported by the file format), e.g. an HDF5 group.

Currently supported file formats are:

HDF5 with file extension ".h5" or ".hdf5"

BAT.bat_rng — Function

bat_rng()

Return a new BAT-compatible random number generator, with a random seed drawn from the system entropy pool.

BAT.bat_sample — Function

bat_sample(
    [rng::AbstractRNG],
    target::BAT.AnySampleable,
    n::BAT.AnyNSamples,
    [algorithm::BAT.AbstractSamplingAlgorithm]
)::DensitySampleVector

Draw n samples from target.

Returns a NamedTuple of the shape

(result = X::DensitySampleVector, ...)

Result properties not listed here are algorithm-specific and are not part of the stable BAT API.

Depending on the type of posterior, n may be of type

Integer: Number of samples

Tuple{Integer,Integer}: Tuple of number of samples per sample source and number of sample sources (e.g. number of MCMC chains). The total number of samples is product(n).

Depending on the type of posterior, the number of samples returned may be somewhat larger or smaller than specified by product(n).

Also depending on the posterior type, the samples may be independent or correlated (e.g. when using MCMC).

Returns a NamedTuple of the shape

(result = X::DensitySampleVector, ...)

Result properties not listed here are algorithm-specific and are not part of the stable BAT API.

Note

Do not add add methods to bat_sample, add methods to bat_sample_impl instead (same arguments).

MCMC Sampling:

function bat_sample(
    rng::AbstractRNG,
    target::AbstractPosteriorDensity,
    n::Union{Integer,Tuple{Integer,Integer}},
    algorithm::MCMCAlgorithm;
    max_nsteps::Integer,
    max_time::Real,
    tuning::AbstractMCMCTuningStrategy,
    init::MCMCInitStrategy,
    burnin::MCMCBurninStrategy,
    convergence::MCMCConvergenceTest,
    strict::Bool = false,
    filter::Bool = true
)

Sample target via Markov chain Monte Carlo (MCMC).

n must be either a tuple (nsteps, nchains) or an integer. nchains specifies the (approximate) number of MCMC steps per chain, nchains the number of MCMC chains. If n is an integer, it is interpreted as nsteps * nchains, and the number of steps and chains are chosen automatically.

BAT.bat_write — Function

bat_write(filename::AbstractString, data)
bat_write(fn_with_subpath::Tuple{AbstractString, AbstractString}, data)

Write data to a file filename, resp. to an internal sub-path of the file (if supported by the file format), e.g. an HDF5 group.

Currently supported file formats are:

HDF5 with file extension ".h5" or ".hdf5"

BAT.logvalgradof — Function

logvalgradof(density::AbstractDensity, v::AbstractVector{<:Real})

Compute the log of the value of a multivariate density function, as well as the gradient of the log-value for the given variate/parameter-values.

Input:

density: density function

v: argument, i.e. variate / parameter-values

Returns a tuple of the log density value and it's gradient.

Note: This function should not be specialized for custom density types!

BAT.logvalof — Function

logvalof(
    density::AbstractDensity,
    v::Any,
    T::Type{:Real} = density_logval_type(v);
    use_bounds::Bool = true,
    strict::Bool = false
)::T

BAT-internal, not part of stable public API.

Evaluates density log-value via logvalof_unchecked.

Throws an exception on any of these conditions:

The variate shape of density (if known) does not match the shape of v.

The return value of logvalof_unchecked is NaN.

The return value of logvalof_unchecked is an equivalent of positive infinity.

Options:

use_bounds: Return an equivalent of negative infinity if v is out of bounds. use_bounds should only be set to false when it is known that the v is within bounds.

strict: Throw an exception if v is out of bounds.

BAT.fft_autocor — Function

fft_autocor(v::AbstractVector{<:Real})
fft_autocor(v::AbstractVectorOfSimilarVectors{<:Real})

Compute the autocorrelation function (ACF) of variate series v, separately for each degree of freedom.

Uses FFT, in contract to StatsBase.autocor.

BAT.fft_autocov — Function

fft_autocov(v::AbstractVector{<:Real})
fft_autocov(v::AbstractVectorOfSimilarVectors{<:Real})

Compute the autocovariance of of variate series v, separately for each degree of freedom.

Uses FFT, in contract to StatsBase.autocov.

BAT.logvalof_unchecked — Function

BAT.logvalof_unchecked(density::AbstractDensity, v::Any)

Compute log of the value of a multivariate density function for the given variate/parameter-values.

Input:

density: density function

v: argument, i.e. variate / parameter-values

Note: If logvalof_unchecked is called with an argument that is out of bounds, the behaviour is undefined. The result for arguments that are not within bounds is implicitly -Inf, but it is the caller's responsibility to handle these cases.

BAT.AHMC — Type

@with_kw struct AHMC <: MCMCAlgorithm metric::HMCMetric = DiagEuclideanMetric() gradient::Module = ForwardDiff integrator::HMCIntegrator = LeapfrogIntegrator() proposal::HMCProposal = NUTS() adaptor::HMCAdaptor = StanHMCAdaptor() end

Keyword arguments

Also see AdvancedHMC.jl for more detailed information on the following HMC related keyword arguments.

Metric

options:

DiagEuclideanMetric()

UnitEuclideanMetric()

DenseEuclideanMetric()

default: metric = DiagEuclideanMetric()

Integrator

options:

LeapfrogIntegrator(step_size::Float64 = 0.0)

JitteredLeapfrogIntegrator(step_size::Float64 = 0.0, jitter_rate::Float64 = 1.0)

TemperedLEapfrogIntegrator(step_size::Float64 = 0.0, tempering_rate::Float64 = 1.05)

default: integrator = LeapfrogIntegrator(step_size::Float64 = 0.0) For step_size = 0.0, the initial stepsize is determined using AdvancedHMC.find_good_stepsize()

Proposal

options:

FixedStepNumber(n_steps::Int64 = 10)

FixedTrajectoryLength(trajectory_length::Float64 = 2.0)

NUTS(sampling::Symbol, nuts::Symbol)

with - sampling = :SliceTS or :MultinomialTS - nuts = :ClassicNoUTurn or :GeneralisedNoUTUrn

default: proposal = NUTS(sampling::Symbol = :MultinomialTS, nuts::Symbol = :ClassicNoUTurn)

Adaptor

options:

MassMatrixAdaptor()

StepSizeAdaptor(step_size::Float64 = 0.8)

NaiveHMCAdaptor(step_size::Float64 = 0.8)

StanHMCAdaptor(step_size::Float64 = 0.8)

default: adaptor = StanHMCAdaptor(step_size::Float64 = 0.8)

BAT.AHMIntegration — Type

AHMIntegration

Adaptive Harmonic Mean Integration algoritm (Caldwell et al.)[https://arxiv.org/abs/1808.08051].

Constructor:

AHMIntegration(;kwargs...)

Optional Parameters/settings (kwargs):

whitening::WhiteningAlgorithm = CholeskyPartialWhitening()

autocorlen::AutocorLenAlgorithm = GeyerAutocorLen()

volumetype::Symbol = :HyperRectangle

max_startingIDs::Int = 10000

max_startingIDs_fraction::Float64 = 2.5

rect_increase::Float64 = 0.1

warning_minstartingids::Int = 16

dotrimming::Bool = true

uncertainty::Vector{Symbol} = [:cov]: List of uncertainty estimation methods to use, first entry will be used for primary result. Valid values:

:cov: Integral uncertainty for integration regions is estimated based on covariance of integrals of subsets of samples in the regions

:ess: Integral uncertainty for integration regions is estimated based on estimated effective number of samples in each region.

BAT.ARPWeighting — Type

struct ARPWeighting{T<:AbstractFloat} <: AbstractWeightingScheme{T}

Sample weighting scheme suitable for accept/reject-based sampling algorithms (e.g. MetropolisHastings). Both accepted and rejected samples become part of the output, with a weight proportional to their original acceptance probability.

Constructors:

ARPWeighting()

ARPWeighting{T}()

BAT.AbstractDensity — Type

AbstractDensity

Subtypes of AbstractDensity must implement the function

BAT.logvalof_unchecked

For likelihood densities this is typically sufficient, since shape, and variate bounds will be inferred from the prior.

Densities with a known variate shape may also implement

ValueShapes.varshape

Densities with known variate bounds may also implement

BAT.var_bounds

Note

The function BAT.var_bounds is not part of the stable public BAT-API, it's name and arguments may change without notice.

BAT.AbstractDensityValue — Type

AbstractDensityValue

Represents a value of a density function at some point.

May either be encoded as the linear value (see LinDVal) or the log of the value (see LinDVal).

Independent of how the value is stored, it's log value can be retrieved using logvalof.

BAT.AbstractMCMCTuningStrategy — Type

AbstractMCMCTuningStrategy

Abstract super-type for MCMC tuning strategies.

BAT.AbstractPosteriorDensity — Type

abstract type AbstractPosteriorDensity <: AbstractDensity end

Abstract super-type for posterior probability densities.

BAT.AbstractWeightingScheme — Type

AbstractWeightingScheme{T<:Real}

Abstract class for sample weighting schemes.

Weight values will have type T.

BAT.AdaptiveMetropolisTuning — Type

AdaptiveMetropolisTuning(...)

Adaptive MCMC tuning strategy for Metropolis-Hastings samplers.

Adapts the proposal function based on the acceptance ratio and covariance of the previous samples.

Fields:

λ: Controls the weight given to new covariance information in adapting the proposal distribution. Defaults to 0.5.

α: Metropolis-Hastings acceptance ratio target, tuning will try to adapt the proposal distribution to bring the acceptance ratio inside this interval. Defaults to IntervalSets.ClosedInterval(0.15, 0.35)

β: Controls how much the spread of the proposal distribution is widened/narrowed depending on the current MH acceptance ratio.

c: Interval for allowed scale/spread of the proposal distribution. Defaults to ClosedInterval(1e-4, 1e2).

r: Reweighting factor. Take accumulated sample statistics of previous tuning cycles into account with a relative weight of r. Set to 0 to completely reset sample statistics between each tuning cycle.

Constructors:

AdaptiveMetropolisTuning(
    λ::Real,
    α::IntervalSets.ClosedInterval{<:Real},
    β::Real,
    c::IntervalSets.ClosedInterval{<:Real},
    r::Real
)

BAT.AnyDensityLike — Constant

BAT.AnyDensityLike = Union{...}

Union of all types that BAT will accept as a probability Density, resp. that convert(AbstractDenstiy, d) supports:

AbstractDensity

Distributions.Distribution

StatsBase.Histogram

BAT.AnyIIDSampleable — Constant

BAT.AnySampleable = Union{...}

Union of all distribution/density-like types that BAT can draw i.i.d. (independent and identically distributed) samples from:

DistLikeDensity

Distributions.Distribution

StatsBase.Histogram

BAT.AnySampleable — Constant

BAT.AnySampleable = Union{...}

Union of all types that BAT can sample from:

AbstractDensity

DensitySampleVector

Distributions.Distribution

StatsBase.Histogram

BAT.AutocorLenAlgorithm — Type

AutocorLenAlgorithm

Abstract type for integrated autocorrelation length estimation algorithms.

BAT.BrooksGelmanConvergence — Type

BrooksGelmanConvergence

Brooks-Gelman $maximum(R^2)$ convergence test.

BAT.CholeskyPartialWhitening — Type

CholeskyPartialWhitening

Whitening transformation based on a Cholesky transformation of the empirical sample covariance matrix.

Only transforms dimensions (degrees of freedom) for which the marginalized distribution aymptotically approaches zero in the positive and negative direction.

BAT.CholeskyWhitening — Type

CholeskyWhitening

Whitening transformation based on a Cholesky transformation of the empirical sample covariance matrix.

BAT.DensitySample — Type

struct DensitySample

A weighted sample drawn according to an statistical density, e.g. a BAT.AbstractDensity.

Fields: * v: Multivariate parameter vector * logd: log of the value of the density at v * weight: Weight of the sample * info: Additional info on the provenance of the sample. Content depends on the sampling algorithm. * aux: Custom user-defined information attatched to the sample.

Constructors:

DensitySample(
    v::AbstractVector{<:Real},
    logd::Real,
    weight::Real,
    info::Any,
    aux::Any
)

Use DensitySampleVector to store vectors of multiple samples with an efficient column-based memory layout.

BAT.DensitySampleVector — Type

DensitySampleVector

Type alias for StructArrays.StructArray{<:DensitySample,...}.

Constructors:

    DensitySampleVector(
        (
            v::AbstractVector{<:AbstractVector{<:Real}},
            logd::AbstractVector{<:Real},
            weight::AbstractVector{<:Real},
            info::AbstractVector{<:Any},
            aux::AbstractVector{<:Any}
        )
    )

    DensitySampleVector(
            v::AbstractVector,
            logval::AbstractVector{<:Real};
            weight::Union{AbstractVector{<:Real}, Symbol} = fill(1, length(eachindex(v))),
            info::AbstractVector = fill(nothing, length(eachindex(v))),
            aux::AbstractVector = fill(nothing, length(eachindex(v)))
        )

For `weight = :RepetitionWeight`, repeated samples will be dropped and the weight of the corresponding sample will be increased.

BAT.DistLikeDensity — Type

DistLikeDensity <: AbstractDensity

A density that implements part of the Distributions.Distribution interface. Such densities are suitable to be used as a priors.

Typically, custom priors should be implemented as subtypes of Distributions.Distribution. BAT will automatically wrap them in a subtype of DistLikeDensity.

Subtypes of DistLikeDensity are required to support more functionality than a AbstractDensity, but less than a Distribution{Multivariate,Continuous}.

A d::Distribution{Multivariate,Continuous} can be converted into (wrapped in) an DistLikeDensity via conv(DistLikeDensity, d).

The following functions must be implemented for subtypes:

BAT.logvalof_unchecked

BAT.var_bounds

ValueShapes.varshape

Distributions.sampler

Statistics.cov

Note

The function BAT.var_bounds is not part of the stable public BAT-API, it's name and arguments may change without notice.

BAT.GelmanRubinConvergence — Type

GelmanRubinConvergence

Gelman-Rubin $\$maximum(R^2)\$$ convergence test.

BAT.GeyerAutocorLen — Type

GeyerAutocorLen <: AutocorLenAlgorithm

Integrated autocorrelation length estimation based on Geyer’s initial monotone sequence criterion

See C. J. Geyer, "Praktical Markov Chain Monte Carlo" (1992) and C. J. Geyer, "Introduction to Markov Chain Monte Carlo" (2011).

Constructor:

GeyerAutocorLen()

The same algorithm is used by STAN (v2.21) and (MCMCChains.jl (v3.0, function ess_rhat))[https://github.com/TuringLang/MCMCChains.jl/blob/v4.0.0/src/ess.jl#L288].

BAT.IIDSampling — Type

IIDSampling

Constructors:

IIDSampling()

Sample via Random.rand. Only supported for posteriors of type Distributions.MultivariateDistribution and BAT.DistLikeDensity.

BAT.IntegrationAlgorithm — Type

IntegrationAlgorithm

Abstract type for integration algorithms.

BAT.LinDVal — Type

LinDVal{T<:Real} <: AbstractDensityValue

Represent the linear value of a statistical density at some point. LinDVal provides means to unambiguously distinguish between linear and log result values of density functions.

Constructor:

LinDVal(d::Real)

Use logvalof to extract the logarithm of the density value from a LinDVal:

    logvalof(LinDVal(d)) == log(d)

See also LogDVal and AbstractDensityValue.

BAT.LogDVal — Type

LogDVal{T<:Real} <: AbstractDensityValue

Represent the logarithm of the value of a statistical density at some point. LogDVal provides means to unambiguously distinguish between linear and log result values of density functions.

Constructor:

LogDVal(logd::Real)

Use logvalof to extract the actual log-density value from a LogDVal:

    logvalof(LogDVal(d)) == d

See also LinDVal and AbstractDensityValue.

BAT.MCMCAlgorithm — Type

abstract type MCMCAlgorithm <: AbstractSamplingAlgorithm end

Note

The details of the MCMCIterator and MCMCAlgorithm API (see below) currently do not form part of the BAT-API, and may change without notice.

The following methods must be defined for subtypes (e.g. for SomeAlgorithm<:MCMCAlgorithm):

(spec::MCMCSpec{<:SomeAlgorithm})(chainid::Integer)::MCMCIterator

To implement a new MCMC algorithm, subtypes of both MCMCAlgorithm and MCMCIterator are required.

BAT.MCMCBurninStrategy — Type

@with_kw struct MCMCBurninStrategy

Defines the MCMC burn-in strategy, specifically the number and length of MCMC tuning/burn-in cycles. Fields:

max_nsamples_per_cycle: Maximum number of MCMC samples to generate per cycle, defaults to 1000. Definition of a sample depends on MCMC algorithm.

max_nsteps_per_cycle: Maximum number of MCMC steps per cycle, defaults to 10000. Definition of a step depends on MCMC algorithm.

max_time_per_cycle: Maximum wall-clock time to spend per cycle, in seconds. Defaults to Inf.

max_ncycles: Maximum number of cycles.

BAT.MCMCInitStrategy — Type

@with_kw struct MCMCInitStrategy

Defines the MCMC chain initialization strategy. Fields:

init_tries_per_chain: Interval that specifies the minimum and maximum number of tries per MCMC chain to find a suitable starting position. Many candidate chains will be created and run for a short time. The chains with the best performance will be selected for tuning/burn-in and MCMC sampling run. Defaults to IntervalSets.ClosedInterval(8, 128).

max_nsamples_init: Maximum number of MCMC samples for each candidate chain. Defaults to 25. Definition of a sample depends on sampling algorithm.

max_nsteps_init: Maximum number of MCMC steps for each candidate chain. Defaults to 250. Definition of a step depends on sampling algorithm.

max_time_init::Int: Maximum wall-clock time to spend per candidate chain, in seconds. Defaults to Inf.

BAT.MCMCIterator — Type

abstract type MCMCIterator end

Represents the current state of a MCMC chain.

Note

The details of the MCMCIterator and MCMCAlgorithm API (see below) currently do not form part of the BAT-API, and may change without notice.

To implement a new MCMC algorithm, subtypes of both MCMCAlgorithm and MCMCIterator are required.

The following methods must be defined for subtypes of MCMCIterator (e.g. SomeMCMCIter<:MCMCIterator):

BAT.mcmc_spec(chain::SomeMCMCIter)::MCMCSpec

BAT.getrng(chain::SomeMCMCIter)::AbstractRNG

BAT.mcmc_info(chain::SomeMCMCIter)::MCMCIteratorInfo

BAT.nsteps(chain::SomeMCMCIter)::Int

BAT.nsamples(chain::SomeMCMCIter)::Int

BAT.current_sample(chain::SomeMCMCIter)::DensitySample

BAT.sample_type(chain::SomeMCMCIter)::Type{<:DensitySample}

BAT.samples_available(chain::SomeMCMCIter, nonzero_weights::Bool = false)::Bool

BAT.get_samples!(appendable, chain::SomeMCMCIter, nonzero_weights::Bool)::typeof(appendable)

BAT.next_cycle!(chain::SomeMCMCIter)::SomeMCMCIter

BAT.mcmc_step!(
    callback::AbstractMCMCCallback,
    chain::SomeMCMCIter
)

The following methods are implemented by default:

algorithm(chain::MCMCIterator)
getposterior(chain::MCMCIterator)
rngseed(chain::MCMCIterator)
DensitySampleVector(chain::MCMCIterator)
mcmc_iterate!(callback, chain::MCMCIterator, ...)
mcmc_iterate!(callbacks, chains::AbstractVector{<:MCMCIterator}, ...)

BAT.MaxDensityLBFGS — Type

MaxDensityLBFGS <: AbstractModeEstimator

Constructors:

MaxDensityLBFGS()

Estimate the mode of a probability density using LBFGS optimization (via Optim.jl). The gradient of the density is computed using forward-mode auto-differentiation (via ForwardDiff.jl).

BAT.MaxDensityNelderMead — Type

MaxDensityNelderMead <: AbstractModeEstimator

Constructors:

MaxDensityNelderMead()

Estimate the mode of a probability density using Nelder-Mead optimization (via Optim.jl).

BAT.MaxDensitySampleSearch — Type

MaxDensitySampleSearch <: AbstractModeEstimator

Constructors:

MaxDensitySampleSearch()

Estimate the mode as the variate with the highest posterior density value within a given set of samples.

BAT.MetropolisHastings — Type

MetropolisHastings

Metropolis-Hastings MCMC sampling algorithm.

Constructors:

MetropolisHastings()
MetropolisHastings(weighting::AbstractWeightingScheme)

BAT.ModeAsDefined — Type

ModeAsDefined <: AbstractModeEstimator

Constructors:

ModeAsDefined()

Get the mode as defined by the density, resp. the underlying distribution (if available), via StatsBase.mode.

BAT.NoWhitening — Type

NoWhitening

No-op whitening transformation, leaves samples unchanged.

BAT.OrderedResampling — Type

OrderedResampling <: AbstractSamplingAlgorithm

Constructors:

OrderedResampling()

Efficiently resamples from a given series of samples, keeping the order of samples.

Can be used to efficiently convert weighted samples into samples with uniform

BAT.PosteriorDensity — Type

PosteriorDensity{
    Li<:AbstractDensity,
    Pr<:DistLikeDensity,
    ...
} <: AbstractPosteriorDensity

A representation of a PosteriorDensity, based a likelihood and prior, a representation of the posterior density is cached internally. The densities be accessed via

getlikelihood(posterior::PosteriorDensity)::Li
getprior(posterior::PosteriorDensity)::Pr

Constructors:

PosteriorDensity(likelihood::AbstractDensity, prior::AbstractDensity)
PosteriorDensity(likelihood::Any, prior::Any)
PosteriorDensity(log_likelihood::Function, prior::Any)

BAT.RandResampling — Type

RandResampling <: AbstractSamplingAlgorithm

Constructors:

RandResampling()

Resamples from a given set of samples.

BAT.RepetitionWeighting — Type

struct RepetitionWeighting{T<:AbstractFloat} <: AbstractWeightingScheme{T}

Sample weighting scheme suitable for sampling algorithms which may repeated samples multiple times in direct succession (e.g. MetropolisHastings). The repeated sample is stored only once, with a weight equal to the number of times it has been repeated (e.g. because a Markov chain has not moved during a sampling step).

Constructors:

RepetitionWeighting()

RepetitionWeighting{T}()

BAT.SokalAutocorLen — Type

SokalAutocorLen <: AutocorLenAlgorithm

Integrated autocorrelation length estimation based on the automated windowing procedure descibed in A. D. Sokal, "Monte Carlo Methods in Statistical Mechanics" (1996)

Constructor:

SokalAutocorLen(;c::Integer = 5)

c: Step size for window search.

Same procedure is used by the emcee Python package (v3.0).

BAT.StatisticalWhitening — Type

StatisticalWhitening

Whitening transformation based statistical whitening.

BAT.WhiteningAlgorithm — Type

WhiteningAlgorithm

Abstract type for sample whitening algorithms.

BAT.AbstractModeEstimator — Type

BAT.AbstractModeEstimator

Abstract type for BAT optimization algorithms.

A typical application for optimization in BAT is mode estimation (see bat_findmode),

BAT.AbstractSamplingAlgorithm — Type

BAT.AbstractSamplingAlgorithm

Abstract type for BAT sampling algorithms. See bat_sample.

BAT.GenericDensity — Type

GenericDensity{F<:Function} <: AbstractDensity

Constructors:

GenericDensity(f)

Turns the density function f into a BAT-compatible [AbstractDensity)(@ref).f(v)must return either a [LogDVal](@ref) (recommended) or a [LinDVal`](@ref).

It must be safe to execute f in parallel on multiple threads and processes.

ValueShapes.totalndof — Function

ValueShapes.totalndof(density::AbstractDensity)::Union{Int,Missing}

Get the number of degrees of freedom of the variates of density. May return missing, if the shape of the variates is not fixed.

ValueShapes.totalndof(density::DistLikeDensity)::Int

Get the number of degrees of freedom of the variates of density. Must not be missing, a DistLikeDensity must have a fixed variate shape.

ValueShapes.varshape — Function

ValueShapes.varshape(
    density::AbstractDensity
)::Union{ValueShapes.AbstractValueShape,Missing}

ValueShapes.varshape(
    density::DistLikeDensity
)::ValueShapes.AbstractValueShape

Get the shapes of the variates of density.

For prior densities, the result must not be missing, but may be nothing if the prior only supports unshaped variate/parameter vectors.